CamiXon.jl

A package for image analysis of backscattered light

Functions

CamiXon.jl

Functions

Index

CamiXon.indices — Method

indices(A::AbstractArray{T,1}, a::T...)  where T

The index (indices) of selected Array elements (default: all elements)

Examples:

A = collect("ahsgh")
indices(A,'h')
1-element Array{Array{Int64,1},1}:
 [2, 5]

indices(A)
4-element Array{Array{Int64,1},1}:
 [1]
 [2, 5]
 [3]
 [4]

A = [1,2,3,4,2]
indices(A,2)
1-element Array{Array{Int64,1},1}:
 [2, 5]

CamiXon.indices_cnt — Method

indices_cnt(A::AbstractArray{T,1}, a::T...)  where T

The number of indices of selected Array elements (default: all elements)

Examples:

A = collect("ahsgh")
indices_cnt(A,'h')
1-element Array{Array{Int64,1},1}:
 2

indices_cnt(A)
4-element Array{Array{Int64,1},1}:
 1
 2
 1
 1

CamiXon.partitions_cnt — Method

partitions_cnt(n::Int,k::Int)

The number of integer partitions of n in k parts

Example:

partitions_cnt(5,2)
 2

CamiXon.partitions_cnt — Method

partitions_cnt(n::Int)

The total number of integer partitions of n

Example:

partitions_cnt(5)
 7

CamiXon.permutations_cnt — Method

permutations_cnt(A::AbstractArray{T,1}; unique = false)  where T

The number of permutations (option: unique permutations) of the elements of a 1D array

Examples:

A = collect("ahsgh")
permutations_cnt(A)
 120

permutations_cnt(A; unique=true)
 60

Index

CamiXon.indices

CamiXon.indices_cnt

CamiXon.partitions_cnt

CamiXon.partitions_cnt

CamiXon.permutations_cnt