CamiXon.jl

A package for image analysis of backscattered light

Table of contents

CamiXon.jl

Table of contents

Functions

Index

Functions

Missing docstring.

Missing docstring for find_all(A::Union{String,AbstractArray{T,1}}, a::T...; count=false) where T. Check Documenter's build log for details.

Missing docstring.

Missing docstring for find_first(A::Union{String,AbstractArray{T,1}}, a::T...; dict=false) where T. Check Documenter's build log for details.

Missing docstring.

Missing docstring for find_last(A::Union{String,AbstractArray{T,1}}, a::T...; dict=false) where T. Check Documenter's build log for details.

CamiXon.canonical_partitions — Function

canonical_partitions(n; header=false, reverse=true)

The canonical partition in integers of the integer n header=true : unit patition included in output

Examples:

canonical_partitions(6; header=true, reverse=false)
6-element Array{Array{Int64,1},1}:
 [6]
 [5, 1]
 [4, 2]
 [3, 3]
 [2, 2, 2]
 [1, 1, 1, 1, 1, 1]

canonical_partitions(6; header=true)
6-element Array{Array{Int64,1},1}:
 [1, 1, 1, 1, 1, 1]
 [2, 2, 2]
 [3, 3]
 [4, 2]
 [5, 1]
 [6]

canonical_partitions(6)
5-element Array{Array{Int64,1},1}:
 [1, 1, 1, 1, 1, 1]
 [2, 2, 2]
 [3, 3]
 [4, 2]
 [5, 1]

CamiXon.integer_partitions — Function

integer_partitions(n [,m]; transpose=false, count=false)

default : The integer partitions of n count=true : The number of integer partitions of n transpose=false/true : for m>0 restricted to partitions with maximum part/length m

definitions: The integer partition of the positive integer n is a nonincreasing sequence of positive integers p1, p2,... pk whose sum is n. The elements of the sequence are called the parts of the partition.

Examples:

integer_partitions(7)
15-element Array{Array{Int64,1},1}:
 [1, 1, 1, 1, 1, 1, 1]
 [2, 2, 2, 1]
 [3, 3, 1]
 [4, 3]
 [5, 2]
 [6, 1]
 [7]
 [2, 2, 1, 1, 1]
 [3, 2, 2]
 [4, 2, 1]
 [5, 1, 1]
 [2, 1, 1, 1, 1, 1]
 [3, 2, 1, 1]
 [4, 1, 1, 1]
 [3, 1, 1, 1, 1]

integer_partitions(7; count=true)
15

integer_partitions(7,4; count=true)
3

integer_partitions(7,4)
3-element Array{Array{Int64,1},1}:
 [4, 3]
 [4, 2, 1]
 [4, 1, 1, 1]

integer_partitions(7,4; transpose=true)
3-element Array{Array{Int64,1},1}:
 [2, 2, 2, 1]
 [3, 2, 1, 1]
 [4, 1, 1, 1]

CamiXon.permutations_cnt — Method

permutations_cnt(A::AbstractArray{T,1}; unique = false)  where T

The number of permutations (option: unique permutations) of the elements of a 1D array

Examples:

A = collect("ahsgh")
permutations_cnt(A)
 120

permutations_cnt(A; unique=true)
 60

Index

CamiXon.canonical_partitions

CamiXon.integer_partitions

CamiXon.permutations_cnt