CamiXon.jl

A package for image analysis of backscattered light

Functions

CamiXon.jl

Functions

Index

CamiXon.indices — Method

function indices(A::AbstractArray{T,N}, a::T...)  where {T,N}

The index (indices) of selected Array elements (default: all elements).

Examples:

A = collect("ahsgh")
indices(A,'h')
1-element Array{Array{Int64,1},1}:
 [2, 5]

indices(A)
4-element Array{Array{Int64,1},1}:
 [1]
 [2, 5]
 [3]
 [4]

CamiXon.indices_cnt — Method

indices_cnt(A::AbstractArray{T,N}, a::T...)  where {T,N}

Count the number of indices of selected Array elements (default: all elements).

Examples:

A = collect("ahsgh")
indices_cnt(A,'h')
1-element Array{Array{Int64,1},1}:
 2

indices_cnt(A)
4-element Array{Array{Int64,1},1}:
 1
 2
 1
 1

CamiXon.partitions_cnt — Method

partitions_cnt(n::Int,k::Int)
The number of integer partitions of n in k parts

CamiXon.partitions_cnt — Method

partitions_cnt(n::Int)
The total number of integer partitions of n

CamiXon.permutations_cnt — Method

permutations_cnt(A::AbstractArray{Any,1}; unique = false)

The number of permutations (option: unique permutations) of the elements of a 1D array.

Examples:

A = collect("ahsgh")
permutations_cnt(A)
 120

permutations_cnt(A; unique=true)
 60

Index

CamiXon.indices

CamiXon.indices_cnt

CamiXon.partitions_cnt

CamiXon.partitions_cnt

CamiXon.permutations_cnt