CamiXon.jl

A package for image analysis of backscattered light

Table of contents

CamiXon.jl

Table of contents

Finite-difference methods

Finite-difference methods

Finite differences

Consider the analytic function $f$ tabulated in forward order (growing index) at $n$ positions on a grid. The finite difference of two adjacent values on a uniform grid is given by the relation

\[\nabla f[n] = f[n]-f[n-1].\]

This called the finite difference in backward difference notation. In this notation the $k^{th}$-order backward difference, which involves $k+1$ points) is defined by a weighted sum over the function values in backward order, $f[n],\ \ldots,\ f[n-k]$,

\[\nabla^k f[n] = f[n] + c_1^kf[n-1] + \cdots + c_k^kf[n-k] = \sum_{j=0}^{k} c_j^kf[n-j].\]

The k+1 coefficients $c_{j}^{k}=(-1)^{j}\binom{k}{j}$ are weight factors (short: weights) defining the summation. Note that $c_{0}^{k}\equiv1$ and $c_{k}^{k}=(-1)^{k}$.

As the function $f$ is tabulated in forward order we change dummy index to write also the summation in forward order (coefficients in backward order)

\[\nabla^k f[n] = \sum_{j=0}^{k} c_{k-j}^kf[n-k+j].\]

Functions:

f_diff_weight(k,j) $\rightarrow c_j^k$

f_diff_weights(k) $\rightarrow \ [c_k^k,\ c_1^k,\ldots,\ c_0^k]$

f_diff_weights_array(kmax) $\rightarrow \ [\ [c_0^0],\ [c_1^1,c_0^1],\ \ldots,\ [c_k^k,\ c_{k-1}^k,\ldots,\ c_0^k] ]$

CamiXon.f_diff_weight — Method

f_diff_weight(k, j)

Weight coefficient

\[c_{j}^{k}=(-1)^{j}\binom{k}{j},\]

of the $k^{th}$-order finite difference operator $\nabla^k$ and corresponding to the function value $f[n-j]$.

Example:

k = 5; j = 3
f_diff_weight(k, j)
 -10

CamiXon.f_diff_weights — Method

f_diff_weights(k)

Weight vector $[c_k^k,\ \ldots,\ c_0^k]$ defining the $k^{th}$-order finite difference operator,

\[\nabla^{k}f[n] =[c_{k}^{k},\thinspace c_{k-1}^{k},\thinspace\ldots,c_{0}^{k}]\left[\begin{array}{c} f[n-k]\\ \vdots\\ f[n] \end{array}\right]=\sum_{j=0}^{k}c_{k-j}^{k}f[n-k+j],\]

for use with the elements of an analytic function, $f$, tabulated in forward order, $f[n-k], ...,f[n]$ (coefficients in backward order).

Example:

k = 3
f_diff_weights(k)
4-element Vector{Int64}:
  1
 -3
  3
 -1

CamiXon.f_diff_weights_array — Method

f_diff_weights_array(kmax)

Collection of weight vectors $[c_k^k,\ \ldots,\ c_0^k]$ defining the finite difference operators $\nabla^0,\ \ldots,\ \nabla^k$.

Example:

kmax = 3
∇ = f_diff_weights_array(kmax)
4-element Vector{Vector{Int64}}:
 [1]
 [-1, 1]
 [1, -2, 1]
 [-1, 3, -3, 1]

Finite difference expansion

Finite-difference calculus builds on the finite-difference expansion, in practice truncated at order $k$,

\[\sum_{p=0}^{\infty}a_{p}\nabla^{p}f[n]=\sum_{p=0}^{k}a_{p}\nabla^{p}f[n]+\cdots.\]

By choosing the order sufficiently high the expansion can be approximated to any desired level of accuracy. A finite-difference expansion of order $k$ is defined by $k+1$ finite-difference expansion coefficients, above denoted by $a_{0},\ \ldots,\ a_{k}$. Note that both the coefficients and the operators appear in forward order. It takes some bookkeeping to rewrite the expansion in the form of a weighted sum over the $k+1$ function values in backward order $f[n],\ldots,f[n-k]$. Substituting the backward-difference expression for $\nabla^k$, we find

\[\sum_{p=0}^{k}a_{p}\nabla^{p}f[n]=\sum_{p=0}^{k}a_{p}\sum_{j=0}^{p}c_{j}^{i}f[n-j]=\sum_{j=0}^{k}\sum_{p=j}^{k}a_{p}c_{j}^{p}f[n-j]=\sum_{j=0}^{k}b_{j}^{k}f[n-j],\]

where the weighted summation is defined by the weights $b_{j}^{k}=\sum_{p=j}^{k}a_{p}c_{j}^{p}$, with $j=0,\,\ldots,\,k$. After changing dummy index to obtain forward order (coefficients in backward order) by the weights are generated by the function f_diff_expansion_weights(coeffs,∇), where ∇ =f_diff_weights_array(kmax).

Functions:

coeffs = $[a_0,\ \ldots,\ a_k]$

f_diff_expansion_weights(coeffs,∇) $\rightarrow [b_k^k,\ ,\ldots,\ b_0^k]$

CamiXon.f_diff_expansion_weights — Method

f_diff_expansion_weights(a, ∇)

Weight vector $[b_k^k,\ ,\ldots,\ b_0^k]$ corresponding to the expansion coefficients $[a_0^k,\ ,\ldots,\ a_k^k]$ of the $k^{th}$-order finite-difference expansion,

\[\sum_{p=0}^{k}a_{p}\nabla^{p}f[n]=\sum_{j=0}^{k}b_{j}^{k}f[n-j]=\sum_{j=0}^{k}b_{k-j}^{k}f[n-k+j],\]

where $f[n-k], ...,f[n]$ are elements of the analytic function $f$ tabulated in forward order. Note the difference in ordering between the finite-difference expansion coefficients, $a_{0},\ \ldots,\ a_{k}$, and the finite-difference expansion weights, $b_k^{k},\ \ldots,\ b_0^{k}$. Furthermore, the coefficients, $a_j$, are independent of $k$, whereas the weights, $b_j^k$, are $k$-dependent.

Example:

k=5
∇ = f_diff_weights_array(k)
a = UnitRange(0,k)
b = f_diff_expansion_weights(a, ∇)
6-element Vector{Int64}:
  15
 -55
  85
 -69
  29
  -5

Lagrangian interpolation

The definition of the finite difference can be rewritten in the form

\[(1-\nabla)f[n+1] = f[n].\]

Inverting this expression we have

\[f[n+1] = (1 - \nabla)^{-1} f[n] = \sum_{p=0}^{\infty}∇^p f[n],\]

\[f[n+2] = (1 - \nabla)^{-2} f[n] = \sum_{p=0}^{\infty}p∇^p f[n],\]

\[\ \ \ \ \ \ \ \ \ \vdots\]

which can be generalized into the form

$f[n+x] = (1 - \nabla)^{-x} f[n] = \sum_{p=0}^{\infty}l_p(x)∇^p f[n],$

where $l_0\equiv 1$ and $l_p(x) = x(x+1)(x+2)\cdots(x+p-1)/p!$ for $p>0$ are the finite-difference expansion coefficients for lagrangian interpolation. These are generated by the function f_diff_expansion_coeffs_interpolation(k,x). Evaluating the finite-difference expansion up to order $k$ we obtain (after changing dummy index to reverse the summation from backward to forward order)

\[f[n+x] =\sum_{p=0}^{k}l_p(x)\nabla^pf[n]= \sum_{j=0}^{k}r_j^k(x)f[n-j]= \sum_{j=0}^{k}r_{k-j}^k(x)f[n-k+j],\]

where the $k+1$ weights $r_j^k(x)= \sum_{p=j}^{k} l_p(x) c_j^p$ are the lagrangian interpolation weights. These are generated for use in backward order by the function f_diff_expansion_weights(coeffs,∇), with ∇ = f_diff_weights_array(k).

Functions:

coeffs = f_diff_expansion_coeffs_interpolation(k,x) $\rightarrow [l_0(x),\ \ldots,\ l_k(x)]$

lagrangian interpolation weights = f_diff_expansion_weights(coeffs,∇) $\rightarrow [r_k^k(x),\ ,\ldots,\ r_0^k(x)]$

CamiXon.f_diff_expansion_coeffs_interpolation — Method

f_diff_expansion_coeffs_interpolation(k::Int, x::T) where T<:Real

Finite-difference expansion coefficient vector $[l_0(x),\ \ldots\,\ l_p(x)]$ defining $k^{th}$-order lagrangian interpolation of the tabulated analytic function $f(n+x)$ at offset position $x$ with respect to the position $n$, with $-k\le x\le 0$,

\[f[n+x] =\sum_{p=0}^{k}l_p(x)\nabla^pf[n],\]

where $l_0\equiv 1$ and $l_p(x) = x(x+1)(x+2)\cdots(x+p-1)/p!$.

Examples:

k=3
∇ = f_diff_weights_array(k)
x=-1
l = f_diff_expansion_coeffs_interpolation(k,x)
r = f_diff_expansion_weights(l, ∇)
println(l,r)
 [1, -2, 1, 0][0, 0, 1, 0]

CamiXon.f_diff_expansion_coeffs_array_interpolation — Method

f_diff_expansion_coeffs_array_interpolation(k::Int, m::Int)

Finite-difference expansion coefficient vector $[l_0(x),\ \ldots\,\ l_p(x)]$ defining $k^{th}$-order lagrangian interpolation of the tabulated analytic function $f(n+x)$ for $k*m+1$ values of $x$.

k = 2; m = 2
o = f_diff_expansion_coeffs_array_interpolation(k,m); println(o)
 [[1.0, 0.0, 0.0], [1.0, -0.5, -0.125], [1.0, -1.0, -0.0], [1.0, -1.5, 0.375], [1.0, -2.0, 1.0]]

m = 1
o = f_diff_expansion_coeffs_array_interpolation(k,m); println(o)
 [[1, 0, 0], [1, -1, 0], [1, -2, 1]]

CamiXon.f_diff_expansion_weights_array — Method

f_diff_expansion_weights_array(n::Int, k::Int, m::Int, l::Vector{Vector{T}}) where T<:Real

Weight factors for $k^{th}$-order finite-difference expansion with finite-difference expansion coefficients $l$ on uniform grid of $(n-1)*m+1$ points.

Example:

n = 7; k = 3; m = 1
l = f_diff_expansion_coeffs_array_interpolation(k, m)
o = f_diff_expansion_weights_array(n, k, m, l); println(o)
 [[1, 0, 0, 0], [1, 0, 0, 0], [1, 0, 0, 0], [1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 1, 0], [0, 0, 0, 1]]

CamiXon.summation_ranges — Method

summation_ranges(n, k, i, m)

Summation ranges for interpolation positions $i = 0,\ 1,\ \ldots,\ (n-1)*m$ as used in $k^{th}$-order lagrangian interpolation of the anaytic function $f$ tabulated in forward order on a uniform grid of $n$ points, f[1], ...,f[n].

Examples:

n = 7; k = 2; m = 1
o = [summation_ranges(n,i,k,m) for i=0:(n-1)*m]; println(o)
 UnitRange{Int64}[1:3, 2:4, 3:5, 4:6, 5:7, 5:7, 5:7]

CamiXon.f_diff_function_sequences — Method

f_diff_function_sequences(f, n::Int, k::Int, m::Int)

Finite-difference interpolation sequences (of $k+1$ function values in forward order) for $k^{th}$-order lagrangian intepolation of the anaytic function $f$ tabulated in forward order at n points, f[1], ...,f[n].

Example:

f = [0,1,2,3,4,5,6]
k = 2; m = 1
o = f_diff_function_sequences(f, k, m); println(o)
 [[0, 1, 2], [1, 2, 3], [2, 3, 4], [3, 4, 5], [4, 5, 6], [4, 5, 6], [4, 5, 6]]

CamiXon.lagrangian_interpolation — Method

lagrangian_interpolation(f::Vector{Float64}, domain::ClosedInterval{Float64}; k=1, i=0)

$k^{th}$-order lagrangian-interpolation with $i$ intermediate point of the analytic function $f$ tabulated in forward order at n points, $f[1],\ \ldots,\ f[n]$.

Example:

f = [0.0,1,2,3,4,5,6,7]
domain = 0.0..1.0
(X,Y) = lagrangian_interpolation(f, domain; k=2, i=1); println((X,Y))
 (0.0:0.07142857142857142:1.0, [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0, 6.5, 7.0])

Lagrangian differentiation

To derive the lagrangian differentiation formulas we differentiate

\[f[n+x] = (1 - \nabla)^{-x} f[n]\]

analytically with respect to $x$.

\[\frac{df}{dx}[n+x]=-ln(1-\nabla)\ (1-\nabla)^{-x}f[n]=\sum_{q=1}^{k}\tfrac{1}{q}∇^{q}\sum_{p=0}^{k}l_{p}(x)∇^{p}f[n].\]

Expanding the r.h.s.\ in powers of $\nabla$ for given values of $n$ and $x$ we obtain an expression of the form

\[\frac{df}{dx}[n+x]=\sum_{p=1}^{k}l_{p}^{\prime}(x)\nabla^{p}f[n],\]

where $l_{p}^{\prime}(x)$ represents the finite-difference expansion coefficients for lagrangian differentiation at position $n+x$. These coefficients are determined numerically by polynomial multiplication after adding the $p=0$ term $l_{0}^{\prime}(x)\equiv 0$. These are generated by the function f_diff_expansion_coeffs_differentiation(k,x). Evaluating the finite-difference expansion up to order $k$ we obtain (after changing dummy index to reverse the summation from backward to forward order)

\[f[n+x]=\sum_{p=0}^{k}l_{p}^{\prime}(x)\nabla^{p}f[n]=\sum_{j=0}^{k}s_{j}^{k}(x)f[n-j]= \sum_{j=0}^{k}s_{k-j}^k(x)f[n-k+j],\]

where the $s_{j}^{k}(x)=\sum_{p=j}^{k}l_{p}^{\prime}(x)c_{j}^{p}$ are the $k$-point lagrangian differentiation weights. These are generated for use in backward order by the function f_diff_expansion_weights(coeffs,∇), with ∇ = f_diff_weights_array(k).

Functions:

coeffs = f_diff_expansion_coeffs_differentiation(k,x) $\rightarrow [l_0(x),\ \ldots,\ l_k(x)]$

lagrangian differentiation weights = f_diff_expansion_weights(coeffs,∇) $\rightarrow [s_k^k(x),\ ,\ldots,\ s_0^k(x)]$

CamiXon.f_diff_expansion_coeffs_differentiation — Method

f_diff_expansion_coeffs_differentiation(k::Int, x::T) where T<:Real

Finite-difference expansion coefficient vector $[l_0(x),\ \ldots\,\ l_p(x)]$ defining $k^{th}$-order lagrangian differentiation of the tabulated analytic function $f(n+x)$ at position $x$.

Example:

k = 2; x = 0
o = f_diff_expansion_coeffs_differentiation(k,x); println(o)
 [0.0, 1.0, -1.5]

CamiXon.f_diff_expansion_coeffs_array_differentiation — Method

f_diff_expansion_coeffs_array_differentiation(k::Int, m::Int)

Finite-difference expansion coefficient vector $[l_0(x),\ \ldots\,\ l_p(x)]$ defining $k^{th}$-order lagrangian differentiation of the tabulated analytic function $f(n+x)$ for $k*m+1$ values of $x$ in the interval $[-k,\ \cdots,\ 0]$ and with $m-1$ intermediate points.

Examples:

k = 2; m = 2
o = f_diff_expansion_coeffs_array_differentiation(k,m); println(o)
 [[0.0, 1.0, 0.5], [0.0, 1.0, 0.0], [0.0, 1.0, -0.5], [0.0, 1.0, -1.0], [0.0, 1.0, -1.5]]

m = 1
o = f_diff_expansion_coeffs_array_differentiation(k,m); println(o)
 [[0.0, 1.0, 0.5], [0.0, 1.0, -0.5], [0.0, 1.0, -1.5]]

CamiXon.lagrangian_differentiation — Method

lagrangian_differentiation(f::Vector{Float64}, domain::ClosedInterval{Float64}; k=1, i=0)

$k^{th}$-order lagrangian differentiation with $i$ intermediate points of the function $f$ tabulated in forward order at n points, $f[1],\ \ldots,\ f[n]$.

Example:

f = [0.0,1,2,3,4,5]
domain = 0.0..5.0
X,Y = lagrangian_differentiation(f, domain; k=2, i = 0); println(X,Y)
 (0.0:1.0:5.0, [1.0, 1.0, 1.0, 1.0, 1.0, 1.0])

Adams-Moulton integration

The Adams-Moulton integration step is given by the expansion

\[y[n+1]-y[n] = -\frac{h \nabla}{ln(1-\nabla)}f[n+1]\]

\[\ \ \ \ \ \ = h\ ( 1 - \frac{1}{2}\nabla - \frac{1}{12}\nabla^2 - \frac{1}{24}\nabla^3 +\cdots)f[n+1]\]

\[\ \ \ \ \ \ = =h (\sum_{k=0}^{\infty}b_k\nabla^k)f[n+1]. `` A closed expression for the Adams-Moulton expansion coefficients ``b_k`` is not available, so these are generated numerically by the function [`f_diff_expansion_coeffs_adams_moulton(k)`](@ref). For the evaluation of the integration step we limit the summation to $k+1$ terms (order ``k``), \]

math y[n+1]-y[n]= h\ (\sum{p=0}^{k}cp\nabla^p)f[n+1]+\cdots,

where ``b_0,\ldots,b_k`` are rational numbers. Extracting the graetest common denominator, ``1/D``, the step becomes

math y[n+1]-y[n]= \frac{h}{D}(\sum{p=0}^{k}bp^{\prime}\nabla^p)f[n+1]+\cdots,

where ``b_0^{\prime},\ldots,b_k^{\prime}`` are integers and ``b_p=b_p^{\prime}/D``. In practice the expansion is restricted to ``k<18`` (as limited by integer overflow). Note that this limit is much higher than values used in calculations (typically up to ``k = 10``).

Function:

`coeffs` = [`f_diff_expansion_coeffs_adams_moulton(k)`](@ref) ``\rightarrow [b_k^k(x),\ ,\ldots,\ b_0^k(x)]``

@docs fdiffexpansioncoeffsadams_moulton(k::Int)

## FITS

FITS stands for 'Flexible Image Transport System'. This is an open standard origionally developed for the astronomy community to store telescope images together with tables of spectral information. Over the years it has developed into a scientific standard - http://fits.gsfc.nasa.gov/iaufwg.

Within CamiXion only the basic FITS functionality is implemented for users not requiring celestal coordinates. The user can create, read and extend .fits files as well as create, edit and delete user-defined metainformation.

A FITS file consists of a sequence of one or more header-data-units (HDUs), each containing a data block preceeded by header records of metainformation.

By the command `f = fits_read(filnam)` we asign a collection of `FITS_HDU` objects from the file `filnam` to the variable `f`.

### FITS - Types

@docs FITSHDU FITSheader FITSdata FITStable FITS_name

### FITS - HDU Methods

@docs fitsinfo(hdu::FITSHDU) parseFITSTABLE(hdu::FITS_HDU)

### FITS - File Methods

@docs castFITSname(filename::String) fitscombine(filnamFirst::String, filnamLast::String; protect=true) fitscopy(filenameA::String, filenameB::String=" "; protect=true) fitscreate(filename::String, data=[]; protect=true) fitsextend(filename::String, dataextend, hdutype="IMAGE") fitsread(filename::String)

### FITS - Key Methods

@docs fitsaddkey(filename::String, hduindex::Int, key::String, val::Real, com::String) fitsdeletekey(filename::String, hduindex::Int, key::String) fitseditkey(filename::String, hduindex::Int, key::String, val::Real, com::String) fitsrenamekey(filename::String, hduindex::Int, keyold::String, keynew::String)

## FORTRAN

@docs FORTRANformat castFORTRANformat(str::String) castFORTRAN_datatype(str::String)

## Plotting

@docs step125(x::Real) select125(x) steps(x::Vector{T} where T<:Real) stepcenters(x::Vector{T} where T<:Real) stepedges(x::Vector{T} where T<:Real) edges(px, Δx=1.0, x0=0.0)

## Search algorithms

@docs findall(A::Union{String,AbstractArray{T,1}}, a::T...; count=false) where T findfirst(A::Union{String,AbstractArray{T,1}}, a::T...; dict=false) where T find_last(A::Union{String,AbstractArray{T,1}}, a::T...; dict=false) where T

## Math

@docs canonicalpartitions(n::Int, m=0; header=true, reverse=true) integerpartitions(n::Int, m=0; transpose=false, count=false) log10characteristicpower(x) log10mantissa(x) polynomderivcoeffs(c,deriv=0) polynom(c::Vector{T}, x::T) where T<:Real polynommultiplicationcoeffs(a::Vector{<:Number}, b::Vector{<:Number}) permutationsunique_count(p::Array{Array{Int64,1},1}, i::Int)

## Index

@index ```