CamiXon.jl

A package for image analysis of backscattered light

Table of contents

CamiXon.jl

Table of contents

FITS

FORTRAN

Search algorithms

Math

Index

FITS

FITS stands for 'Flexible Image Transport System'. This is an open standard origionally developed for the astronomy community to store telescope images together with tables of spectral information. Over the years it has developed into a scientific standard - http://fits.gsfc.nasa.gov/iaufwg.

Within CamiXion only the basic FITS functionality is implemented for users not requiring celestal coordinates. The user can create, read and extend .fits files as well as create, edit and delete user-defined metainformation.

A FITS file consists of a sequence of one or more header-data-units (HDUs), each containing a data block preceeded by header records of metainformation.

By the command f = fits_read(filnam) we asign a collection of FITS_HDU objects from the file filnam to the variable f.

CamiXon.FITS_HDU — Type

FITS_HDU

Object to hold a single "Header-Data Unit" (HDU).

The fields are

.filename::String: name of the corresponding FITS file

.hduindex::Int: identifier (a file may contain more than one HDU)

.header::T, where T=FITS_header: the header object

.dataobject::V, where V=FITS_data: the data object

CamiXon.FITS_header — Type

FITS_header

Object to hold the header information of a FITS_HDU.

The fields are:

.hduindex::Int: identifier (a file may contain more than one HDU)

.records::Array{String,1}: the header formated as an array of strings of 80 ASCII characters

.keys::Array{String,1}: keys[i] - key corresponding to records[i] (record of index i)

.values::Array{Any,1}: value[i] - corresponding to records[i]

.comments: comments[i] - comment corresponding to records[i]

.dict::Dict{String,Any}: Dictionary key[i] => value[i]

.maps::Dict{String,Int}: Dictionary key[i] => i

CamiXon.FITS_data — Type

FITS_data

Object to hold the data of the FITS_HDU of given hduindex and hdutype.

The fields are:

.hduindex::Int: identifier (a file may contain more than one HDU)

.hdutype::String: accepted types are 'PRIMARY', 'IMAGE' and 'TABLE'

.data::Any: in the from appropriate for the hdutype

CamiXon.FITS_table — Type

FITS_table

Object to hold the data of a TABLE HDU (a FITS_HDU for ASCII tables). It contains the data in the form of records (rows) of ASCII strings.

The fields are:

.hduindex::Int: identifier (a file may contain more than one HDU)

.rows::Array{String,1}: the table formated as an array of rows of ASCII strings

CamiXon.parse_FITS_TABLE — Method

parse_FITS_TABLE(FITS_HDU)

Parse FITS_TABLE into a Vector of its columns for further processing by the user.

Example:

f = fits_create("minimal.fits";protect=false)
fits_info(f[1])

 File: minimal.fits
 HDU: 1
 DataType: Any
 Datasize: (0,)

 Metainformation:
 SIMPLE  =                    T / file does conform to FITS standard             
 NAXIS   =                    0 / number of data axes                            
 EXTEND  =                    T / FITS dataset may contain extensions            
 COMMENT    Basic FITS file     / http://fits.gsfc.nasa.gov/iaufwg               
 END

CamiXon.FITS_name — Type

FITS_name

FITS object to decompose the names of .fits files.

The fields are:

.name::String: for filename 'p#.fits' this is 'p#.fits'

.prefix::String: for filename 'p#.fits' this is 'p'

.numerator::String: for filename 'p#.fits' this is '#', a serial number (e.g., '3') or a range (e.g., '3-7')

.extension::String: for filename 'p#.fits' this is '.fits'.

CamiXon.cast_FITS_name — Method

cast_FITS_name(filename::String)

Decompose the FITS filename 'filnam.fits' into its name, prefix, numerator and extension.

Examples:

strExample = "T23.01.fits"
f = analyze_FITS_name(strExample)
FITS_name("T23.01", "T23.", "01", ".fits")

f.name, f.prefix, f.numerator, f.extension
("T23.01", "T23.", "01", ".fits")

CamiXon.fits_info — Method

fits_info(FITS_HDU; printformat=true)

Print metafinformation of given FITS_HDU

Key:

printformat::Bool: output formatted by function print

Example:

strExample = "minimal.fits"
f = fits_create(strExample; protect=false)
fits_info(f[1])

  File: minimal.fits
  HDU: 1
  DataType: Any
  Datasize: (0,)

  Metainformation:
  SIMPLE  =                    T / file does conform to FITS standard             
  NAXIS   =                    0 / number of data axes                            
  EXTEND  =                    T / FITS dataset may contain extensions            
  COMMENT    Basic FITS file     / http://fits.gsfc.nasa.gov/iaufwg               
  END

CamiXon.fits_create — Function

fits_create(filename [, data [; protect=true]])

Create FITS file of given filename [, optional data block [, default overwrite protection]] and return Array of HDUs. Key:

protect::Bool: overwrite protection

Examples:

strExample = "minimal.fits"
f = fits_create(strExample;protect=false)
a = f[1].dataobject.data
b = f[1].header.keys
println(a);println(b)
  Any[]
  ["SIMPLE", "NAXIS", "EXTEND", "COMMENT", "END"]

strExample = "example.fits"
data = [0x0000043e, 0x0000040c, 0x0000041f]
f = fits_create(strExample, data; protect=false)
a = f[1].dataobject.data
b = f[1].header.keys
println(a);println(b)
  UInt32[0x0000043e, 0x0000040c, 0x0000041f]
  ["SIMPLE", "BITPIX", "NAXIS", "NAXIS1", "BZERO", "BSCALE", "EXTEND", "COMMENT", "END"]

CamiXon.fits_read — Method

fits_read(filename)

Read FITS file and return Array of FITS_HDUs

Example:

strExample = "minimal.fits"
fits_create(strExample;protect=false)
f = fits_read(strExample)
f[1].dataobject.data
  Any[]

rm(strExample); f = nothing

CamiXon.fits_extend — Function

fits_extend(filename, data_extend, hdutype="IMAGE")

Extend the FITS file of given filename with the data of hdutype from data_extend and return Array of HDUs.

Examples:

strExample = "test_example.fits"
data = [0x0000043e, 0x0000040c, 0x0000041f]
f = fits_create(strExample, data, protect=false)
table1 = Float16[1.01E-6,2.0E-6,3.0E-6,4.0E-6,5.0E-6]
table2 = [0x0000043e, 0x0000040c, 0x0000041f, 0x0000042e, 0x0000042f]
table3 = [1.23,2.12,3.,4.,5.]
table4 = ['a','b','c','d','e']
table5 = ["a","bb","ccc","dddd","ABCeeaeeEEEEEEEEEEEE"]
data = [table1,table2,table3,table4,table5]
f = fits_extend(strExample, data, "TABLE")
f[2].dataobject.data
  5-element Vector{String}:
   "1.0e-6 1086 1.23 a a                    "
   "2.0e-6 1036 2.12 b bb                   "
   "3.0e-6 1055 3.0  c ccc                  "
   "4.0e-6 1070 4.0  d dddd                 "
   "5.0e-6 1071 5.0  e ABCeeaeeEEEEEEEEEEEE "

rm(strExample); f = nothing; data =nothing

CamiXon.fits_copy — Function

fits_copy(filenameA [, filenameB="" [; protect=true]])

Copy "filenameA" to "filenameB" (with mandatory ".fits" extension) Key:

protect::Bool: overwrite protection

Examples:

fits_copy("T01.fits")
  'T01.fits' was saved as 'T01 - Copy.fits'

fits_copy("T01.fits", "T01a.fits")
  filename (T01a.fits) in use (overwrite protected)

fits_copy("T01.fits", "T01a.fits"; protect=false)
  'T01.fits' was saved as 'T01a.fits'

Missing docstring.

Missing docstring for fits_add_key(filename::String, hduindex::Int, key::String, val::Real, com::String). Check Documenter's build log for details.

CamiXon.fits_edit_key — Method

fits_edit_key(filename, hduindex, key, value, comment)

Edit a header record of given 'key, value and comment' to 'HDU[hduindex]' of file with name 'filename'

Example:

filnam="minimal.fits"
fits_create(filnam;protect=false)
fits_add_key(filnam, 1, "EXTEND12", true, "FITS dataset may contain extension")
  'EXTEND12': key added; new record: 'EXTEND12=                    T / FITS dataset may contain extension             '

fits_edit_key(filnam, 1, "EXTEND12", true, "FITS dataset new comment")
  'EXTEND12': key edited; new record: 'EXTEND12=                    F / FITS dataset new comment                       '

CamiXon.fits_delete_key — Method

fits_delete_key(filename, hduindex, key)

Delete a header record of given key, value and comment to FITS_HDU[hduindex] of file with name 'filename'

Examples:

filnam="minimal.fits"
fits_create(filnam;protect=false)
fits_add_key(filnam, 1, "EXTEND123", true, "FITS dataset may contain extension")
 'EXTEND123' key truncated at 8 characters (FITS standard)
 'EXTEND12': key added; new record: 'EXTEND12=                    T / FITS dataset may contain extension             '

fits_add_key(filnam, 1, "EXTEND12", true, "FITS dataset may contain extension")
 'EXTEND12': key in use (use different key name or edit key)

fits_delete_key(filnam, 1, "EXTEND12")
 'EXTEND12': key deleted

fits_delete_key(filnam, 1, "EXTEND12")
 'EXTEND12': cannot be deleted (key not found)

fits_delete_key(filnam, 1, "NAXIS")
 'NAXIS': cannot be deleted (key protected under FITS standard)

FORTRAN

CamiXon.FORTRAN_format — Type

FORTRAN_format

Object to hold a FORTRAN format specifier decomposed into its fields. Accepted datatype specifiers are: Aw, Iw, Fw.d, Ew.d, Dw.d. Accepted output formating specifiers are: Aw, Iw.m, Bw.m, Ow.m, Zw.m, Fw.d, Ew.dEe, ENw.d, ESw.d, Gw.dEe, Dw.dEe. Notation: 'w' - width, 'm' (optional) - minimum number of digits, 'd' - number of digits to right of decimal, 'e' - number of digits in exponent; 'N'/'S' (optional) - switch for engineering/scientific formating of the 'E' type.

The fields are:

Type::String: primary FORTRAN datatype

TypeChar::Char: primary FORTRAN datatype character

EngSci::Union{Char,Nothing}: secundary datatype character (N for engineering, S for scientific)

width::Int: width of numeric field

nmin::Int: minimum number of digits displayed

ndec::Int: number of digits to right of decimal

nexp::Int: number of digits in exponent

CamiXon.cast_FORTRAN_format — Method

cast_FORTRAN_format(format::String)

Decompose the format specifier 'format' into its fields and cast this into the FORTRAN_format object. Allowed format specifiers are of the types: Aw, Iw.m, Bw.m, Ow.m, Zw.m, Fw.d, Ew.dEe, ENw.d, ESw.d, Gw.dEe, Dw.dEe, with: 'w' - width, 'm' (optional) - minimum number of digits, 'd' - number of digits to right of decimal, 'e' - number of digits in exponent; 'N'/'S' (optional) - switch for engineering/scientific formating of the 'E' type.

Examples:

f = cast_FORTRAN_format("I10")
FORTRAN_format("Iw", 'I', nothing, 10, 0, 0, 0)

f = cast_FORTRAN_format("I10.12")
FORTRAN_format("Iw.m", 'I', nothing, 10, 12, 0, 0)

f = cast_FORTRAN_format("E10.5E3")
FORTRAN_format("Ew.dEe", 'E', nothing, 10, 0, 5, 3)

t.Type, t.TypeChar, t.EngSci, t.width, t.nmin, t.ndec, t.nexp
("Ew.dEe", 'E', nothing, 10, 0, 5, 3)

CamiXon.cast_FORTRAN_datatype — Method

cast_FORTRAN_datatype(format::String)

Decompose the format specifier 'format' into its fields and cast this into the FORTRAN_format object. Allowed format specifiers are of the types: Aw, Iw, Fw.d, Ew.d, Dw.d, where: 'w' - width, 'd' - number of digits to right of decimal point.

Examples:

f = cast_FORTRAN_datatype("I10")
FORTRAN_format("Iw", 'I', nothing, 10, 0, 0, 0)

f = cast_FORTRAN_datatypet("F10.4")
FORTRAN_format("Fw.d", 'F', nothing, 10, 0, 4, 0)

f = cast_FORTRAN_datatype("E10.5")
FORTRAN_format("Ew.d", 'E', nothing, 10, 0, 5, 0)

t.Type, t.TypeChar, t.EngSci, t.width, t.nmin, t.ndec, t.nexp
("Ew.d", 'E', nothing, 10, 0, 5, 0)

Search algorithms

CamiXon.find_all — Method

find_all(A [,a...]; count=false)

A: string/array of elements of the same type

default : Array containing the index (indices) of selected elements of A (default: all elements)

count=true: The number of indices found for selected elements of A (default: all elements)

Examples:

A = [:📑,:📌,:📢,:📌,:📞]
B = [1,2,3,2,5]
str = "aβcβd";
find_all(A) == find_all(B) == find_all(str)
true

find_all(A,:📌)
1-element Array{Array{Int64,1},1}:
 [2, 4]

find_all(str)
4-element Array{Array{Int64,1},1}:
 [1]
 [2, 4]
 [3]
 [5]

find_all(A; count=true)
4-element Array{Int64,1}:
 1
 2
 1
 1

str = "📑📌📢📌📞"
find_all(str,'📌')
1-element Array{Array{Int64,1},1}:
 [2, 4]

CamiXon.find_first — Method

find_first(A [,a...]; dict=false)

The first index of selected Array element

A: string/array of elements of the same type

default : Array containing the first index (indices) of selected elements of A (default: all elements)

dict=true: Dict for the first index (indices) of selected elements of A (default: all elements)

Examples:

A = [:📑,:📌,:📢,:📌,:📞]
B = [1,2,3,2,5]
str = "aβcβd";

find_first(A) == find_first(B) == find_first(str)
true

find_first(A,:📌)
1-element Array{Array{Int64,1},1}:
 2

find_last(A,:📌; dict=true)
1-element Array{Pair{Symbol,Int64},1}:
 :📌 => 2

find_last(A; dict=true)
4-element Array{Pair{Symbol,Int64},1}:
 :📑 => 1
 :📌 => 2
 :📢 => 3
 :📞 => 5

find_first(str)
4-element Array{Int64,1}:
 1
 2
 3
 5

CamiXon.find_last — Method

find_last(A [,a...]; dict=false)

The last index of selected Array element

A: string/array of elements of the same type

default : Array containing the lasst index (indices) of selected elements of A (default: all elements)

dict=true: Dict for the lasst index (indices) of selected elements of A (default: all elements)

Examples:

A = [:📑,:📌,:📢,:📌,:📞]
B = [1,2,3,2,5]
str = "aβcβd";
find_last(A) == find_first(B) == find_first(str)
true

find_last(A,:📌)
1-element Array{Array{Int64,1},1}:
 4

find_last(A,:📌; dict=true)
1-element Array{Pair{Symbol,Int64},1}:
 :📌 => 4

find_last(A; dict=true)
4-element Array{Pair{Symbol,Int64},1}:
 :📑 => 1
 :📌 => 4
 :📢 => 3
 :📞 => 5

find_last(str)
4-element Array{Int64,1}:
 1
 4
 3
 5

Math

CamiXon.canonical_partitions — Function

canonical_partitions(n; header=false, reverse=true)

The canonical partition in integers of the integer n

header=true : unit patition included in output

Examples:

canonical_partitions(6; header=true, reverse=false)
6-element Array{Array{Int64,1},1}:
 [6]
 [5, 1]
 [4, 2]
 [3, 3]
 [2, 2, 2]
 [1, 1, 1, 1, 1, 1]

canonical_partitions(6; header=true)
6-element Array{Array{Int64,1},1}:
 [1, 1, 1, 1, 1, 1]
 [2, 2, 2]
 [3, 3]
 [4, 2]
 [5, 1]
 [6]

canonical_partitions(6)
5-element Array{Array{Int64,1},1}:
 [1, 1, 1, 1, 1, 1]
 [2, 2, 2]
 [3, 3]
 [4, 2]
 [5, 1]

CamiXon.integer_partitions — Function

integer_partitions(n [,m]; transpose=false, count=false)

default : The integer partitions of n

count=true : The number of integer partitions of n

transpose=false/true : for m>0 restricted to partitions with maximum part/length m

definitions:

The integer partition of the positive integer n is a nonincreasing sequence of positive integers p1, p2,... pk whose sum is n.

The elements of the sequence are called the parts of the partition.

Examples:

integer_partitions(7)
15-element Array{Array{Int64,1},1}:
 [1, 1, 1, 1, 1, 1, 1]
 [2, 2, 2, 1]
 [3, 3, 1]
 [4, 3]
 [5, 2]
 [6, 1]
 [7]
 [2, 2, 1, 1, 1]
 [3, 2, 2]
 [4, 2, 1]
 [5, 1, 1]
 [2, 1, 1, 1, 1, 1]
 [3, 2, 1, 1]
 [4, 1, 1, 1]
 [3, 1, 1, 1, 1]

integer_partitions(7; count=true)
15

integer_partitions(7,4; count=true)
3

integer_partitions(7,4)
3-element Array{Array{Int64,1},1}:
 [4, 3]
 [4, 2, 1]
 [4, 1, 1, 1]

integer_partitions(7,4; transpose=true)
3-element Array{Array{Int64,1},1}:
 [2, 2, 2, 1]
 [3, 2, 1, 1]
 [4, 1, 1, 1]

Index

CamiXon.FITS_HDU

CamiXon.FITS_data

CamiXon.FITS_header

CamiXon.FITS_name

CamiXon.FITS_table

CamiXon.FORTRAN_format

CamiXon.canonical_partitions

CamiXon.cast_FITS_name

CamiXon.cast_FORTRAN_datatype

CamiXon.cast_FORTRAN_format

CamiXon.find_all

CamiXon.find_first

CamiXon.find_last

CamiXon.fits_copy

CamiXon.fits_create

CamiXon.fits_delete_key

CamiXon.fits_edit_key

CamiXon.fits_extend

CamiXon.fits_info

CamiXon.fits_read

CamiXon.integer_partitions

CamiXon.parse_FITS_TABLE