CamiXon

Here's an equation:

\[\frac{n!}{k!(n - k)!} = \binom{n}{k}\]

This is the binomial coefficient.

To write a system of equations, use the aligned environment:

\[\begin{aligned} \nabla\cdot\mathbf{E} &= 4 \pi \rho \\ \nabla\cdot\mathbf{B} &= 0 \\ \nabla\times\mathbf{E} &= - \frac{1}{c} \frac{\partial\mathbf{B}}{\partial t} \\ \nabla\times\mathbf{B} &= - \frac{1}{c} \left(4 \pi \mathbf{J} + \frac{\partial\mathbf{E}}{\partial t} \right) \end{aligned}\]

These are Maxwell's equations.

CamiXon.get_indices — Method

function get_indices(A::AbstractArray{T,N}, a::T...)  where {T,N}

Find the index (indices) of selected Array elements (default: all elements).

Examples:

A = collect("ahsgh")
get_indices(A,'h')
1-element Array{Array{Int64,1},1}:
 [2, 5]

get_indices(A)
4-element Array{Array{Int64,1},1}:
 [1]
 [2, 5]
 [3]
 [4]

CamiXon.get_indices_count — Method

get_indices_count(A::AbstractArray{T,N}, a::T...)  where {T,N}

Count the number of indices of selected Array elements (default: all elements).

Examples:

A = collect("ahsgh")
get_indices_count(A,'h')
1-element Array{Array{Int64,1},1}:
 2

get_indices_count(A)
4-element Array{Array{Int64,1},1}:
 1
 2
 1
 1

CamiXon.get_permutation_count — Method

get_permutation_count(A::AbstractArray{T,N}; unique = false)  where {T,N}

Count the number of permutations (option: unique permutations).

Examples:

A = collect("ahsgh")
permutation_count(A)
 120

permutation_count(A; unique=true)
 60

CamiXon.get_indices

CamiXon.get_indices_count

CamiXon.get_permutation_count