CamiXon.jl

A package for image analysis of backscattered light

Table of contents

CamiXon.jl

Table of contents

File manipulation

Search algorithms

Math

Index

File manipulation

CamiXon.decompose_filnam — Method

decompose_filnam(str)

Decompose filename into its name (and, if present, extension, prefix and numerator).

Examples:

strExample = "T23.01.fits"

dict = decompose_filnam(strExample)
Dict{String,String} with 4 entries:
  "Extension" => ".fits"
  "Numerator" => "01"
  "Prefix"    => "T23."
  "Name"      => "T23.01"

get(dict,"Numerator","Numerator: Key absent")
"01"

get(dict,"Wild","Key absent")
"Key absent"

CamiXon.fits_combine — Method

fits_combine(str1, str2 [; info=false])

Combine a series of .fits files into a single .fits file.

Example:

fits_combine("T01.fits", "T22.fits"; info=false)
T01-T22.FITS: file was created (for more information set info=true)

CamiXon.fits_copy — Function

fits_copy(filnam [, filnam2="" [; protect=true]])

Copy "filnam.fits" to "filnam2.fits"

Examples:

fits_copy("T01.fits")
T01.fits was saved as T01 - Copy.fits

fits_copy("T01.fits", "T01a.fits")
"T01a.fits: filname in use (overwrite protected)"

fits_copy("T01.fits", "T01a.fits"; protect=false)
T01.fits was saved as T01a.fits

CamiXon.fits_info — Method

fits_info(filnam [; info=false])

Metainformation of 'filnam.fits'

Example:

fits_info("T01.fits"; info=false)
T01.fits: file was found (for more information set info=true)

CamiXon.fits_key_create — Method

fits_key_create(filnam, key, value, comment)

create FITS key record

Example:

filnam = "T01.fits"
fits_key_create(filnam,"Test",3,"this is a test")
("TEST", 3, "this is a test")

CamiXon.fits_key_delete — Method

fits_key_delete(filnam, key)

edit FITS key record

Example:

filnam = "T01.fits"
fits_key_info(filnam,"test")
"key 'TEST' not in use"

fits_key_create(filnam,"Test",3,"this is a test")
("TEST", 3, "this is a test")

fits_key_delete(filnam,"test")
"TEST: key succesfully deleted"

fits_key_info(filnam,"test1")
"key 'TEST' not in use"

CamiXon.fits_key_edit — Method

fits_key_edit(filnam, key, value, comment)

edit FITS key record

Example:

filnam = "T01.fits"
fits_key_edit(filnam, "USERTXT1", 3, "my text")
("USERTXT1", 3, "my text")

CamiXon.fits_key_info — Method

fits_key_info(filnam::String, key::String)

show FITS key record

Example:

filnam = "T01.fits"
fits_key_info(filnam,"NAXIS1")
("NAXIS1", 512, "length of data axis 1")

CamiXon.fits_key_rename — Method

fits_key_rename(filnam, key, keynew)

rename FITS key

Examples:

filnam = "T01.fits"
fits_key_info(filnam,"test1")
"key 'TEST1' not in use"

fits_key_create(filnam,"Test",3,"this is a test")
("TEST", 3, "this is a test")

fits_key_rename(filnam,"test","test1")
("TEST1", 3, "this is a test")

fits_key_info(filnam,"test1")
("TEST1", 3, "this is a test")

Search algorithms

CamiXon.find_all — Method

find_all(A [,a...]; count=false)

A: string/array of elements of the same type

default : Array containing the index (indices) of selected elements of A (default: all elements)

count=true: The number of indices found for selected elements of A (default: all elements)

Examples:

A = [:📑,:📌,:📢,:📌,:📞]
B = [1,2,3,2,5]
str = "aβcβd";
find_all(A) == find_all(B) == find_all(str)
true

find_all(A,:📌)
1-element Array{Array{Int64,1},1}:
 [2, 4]

find_all(str)
4-element Array{Array{Int64,1},1}:
 [1]
 [2, 4]
 [3]
 [5]

find_all(A; count=true)
4-element Array{Int64,1}:
 1
 2
 1
 1

str = "📑📌📢📌📞"
find_all(str,'📌')
1-element Array{Array{Int64,1},1}:
 [2, 4]

CamiXon.find_first — Method

find_first(A [,a...]; dict=false)

The first index of selected Array element

A: string/array of elements of the same type

default : Array containing the first index (indices) of selected elements of A (default: all elements)

dict=true: Dict for the first index (indices) of selected elements of A (default: all elements)

Examples:

A = [:📑,:📌,:📢,:📌,:📞]
B = [1,2,3,2,5]
str = "aβcβd";

find_first(A) == find_first(B) == find_first(str)
true

find_first(A,:📌)
1-element Array{Array{Int64,1},1}:
 2

find_last(A,:📌; dict=true)
1-element Array{Pair{Symbol,Int64},1}:
 :📌 => 2

find_last(A; dict=true)
4-element Array{Pair{Symbol,Int64},1}:
 :📑 => 1
 :📌 => 2
 :📢 => 3
 :📞 => 5

find_first(str)
4-element Array{Int64,1}:
 1
 2
 3
 5

CamiXon.find_last — Method

find_last(A [,a...]; dict=false)

The last index of selected Array element

A: string/array of elements of the same type

default : Array containing the lasst index (indices) of selected elements of A (default: all elements)

dict=true: Dict for the lasst index (indices) of selected elements of A (default: all elements)

Examples:

A = [:📑,:📌,:📢,:📌,:📞]
B = [1,2,3,2,5]
str = "aβcβd";
find_last(A) == find_first(B) == find_first(str)
true

find_last(A,:📌)
1-element Array{Array{Int64,1},1}:
 4

find_last(A,:📌; dict=true)
1-element Array{Pair{Symbol,Int64},1}:
 :📌 => 4

find_last(A; dict=true)
4-element Array{Pair{Symbol,Int64},1}:
 :📑 => 1
 :📌 => 4
 :📢 => 3
 :📞 => 5

find_last(str)
4-element Array{Int64,1}:
 1
 4
 3
 5

Math

CamiXon.canonical_partitions — Function

canonical_partitions(n; header=false, reverse=true)

The canonical partition in integers of the integer n

header=true : unit patition included in output

Examples:

canonical_partitions(6; header=true, reverse=false)
6-element Array{Array{Int64,1},1}:
 [6]
 [5, 1]
 [4, 2]
 [3, 3]
 [2, 2, 2]
 [1, 1, 1, 1, 1, 1]

canonical_partitions(6; header=true)
6-element Array{Array{Int64,1},1}:
 [1, 1, 1, 1, 1, 1]
 [2, 2, 2]
 [3, 3]
 [4, 2]
 [5, 1]
 [6]

canonical_partitions(6)
5-element Array{Array{Int64,1},1}:
 [1, 1, 1, 1, 1, 1]
 [2, 2, 2]
 [3, 3]
 [4, 2]
 [5, 1]

CamiXon.integer_partitions — Function

integer_partitions(n [,m]; transpose=false, count=false)

default : The integer partitions of n

count=true : The number of integer partitions of n

transpose=false/true : for m>0 restricted to partitions with maximum part/length m

definitions:

The integer partition of the positive integer n is a nonincreasing sequence of positive integers p1, p2,... pk whose sum is n.

The elements of the sequence are called the parts of the partition.

Examples:

integer_partitions(7)
15-element Array{Array{Int64,1},1}:
 [1, 1, 1, 1, 1, 1, 1]
 [2, 2, 2, 1]
 [3, 3, 1]
 [4, 3]
 [5, 2]
 [6, 1]
 [7]
 [2, 2, 1, 1, 1]
 [3, 2, 2]
 [4, 2, 1]
 [5, 1, 1]
 [2, 1, 1, 1, 1, 1]
 [3, 2, 1, 1]
 [4, 1, 1, 1]
 [3, 1, 1, 1, 1]

integer_partitions(7; count=true)
15

integer_partitions(7,4; count=true)
3

integer_partitions(7,4)
3-element Array{Array{Int64,1},1}:
 [4, 3]
 [4, 2, 1]
 [4, 1, 1, 1]

integer_partitions(7,4; transpose=true)
3-element Array{Array{Int64,1},1}:
 [2, 2, 2, 1]
 [3, 2, 1, 1]
 [4, 1, 1, 1]

Index

CamiXon.canonical_partitions

CamiXon.decompose_filnam

CamiXon.find_all

CamiXon.find_first

CamiXon.find_last

CamiXon.fits_combine

CamiXon.fits_copy

CamiXon.fits_info

CamiXon.fits_key_create

CamiXon.fits_key_delete

CamiXon.fits_key_edit

CamiXon.fits_key_info

CamiXon.fits_key_rename

CamiXon.integer_partitions