CamiXon.jl

A package for image analysis of backscattered light

Table of contents

CamiXon.jl

Table of contents

Finite-difference methods

FITS

FORTRAN

Plotting

Search algorithms

Math

Index

Finite-difference methods

Finite differences

Consider the analytic function $f$ tabulated in forward order (growing index) at $n$ positions on a grid. The finite difference of two adjacent values on a uniform grid is given by the relation

\[\nabla f[n] = f[n]-f[n-1].\]

This called the finite difference in backward difference notation. In this notation the $k^{th}$-order backward difference, which involves $k+1$ points) is defined by a weighted sum over the function values in backward order, $f[n],\ \ldots,\ f[n-k]$,

\[\nabla^k f[n] = f[n] + c_1^kf[n-1] + \cdots + c_k^kf[n-k] = \sum_{j=0}^{k} c_j^kf[n-j].\]

The k+1 coefficients $c_{j}^{k}=(-1)^{j}\binom{k}{j}$ are weight factors (short: weights) defining the summation. Note that $c_{0}^{k}\equiv1$ and $c_{k}^{k}=(-1)^{k}$.

As the function $f$ is tabulated in forward order we change dummy index to write also the summation in forward order (coefficients in backward order)

\[\nabla^k f[n] = \sum_{j=0}^{k} c_{k-j}^kf[n-k+j].\]

Functions:

f_diff_weight(k,j) $\rightarrow c_j^k$

f_diff_weights(k) $\rightarrow \ [c_k^k,\ c_1^k,\ldots,\ c_0^k]$

f_diff_weights_array(kmax) $\rightarrow \ [\ [c_0^0],\ [c_1^1,c_0^1],\ \ldots,\ [c_k^k,\ c_{k-1}^k,\ldots,\ c_0^k] ]$

CamiXon.f_diff_weight — Method

f_diff_weight(k, j)

Weight coefficient

\[c_{j}^{k}=(-1)^{j}\binom{k}{j},\]

of the $k^{th}$-order finite difference operator $\nabla^k$ and corresponding to the function value $f[n-j]$.

Example:

k = 5; j = 3
f_diff_weight(k, j)
 -10

CamiXon.f_diff_weights — Method

f_diff_weights(k)

Weight vector $[c_k^k,\ \ldots,\ c_0^k]$ defining the $k^{th}$-order finite difference operator,

\[\nabla^{k}f[n] =[c_{k}^{k},\thinspace c_{k-1}^{k},\thinspace\ldots,c_{0}^{k}]\left[\begin{array}{c} f[n-k]\\ \vdots\\ f[n] \end{array}\right]=\sum_{j=0}^{k}c_{k-j}^{k}f[n-k+j],\]

for use with the elements of an analytic function, $f$, tabulated in forward order, $f[n-k], ...,f[n]$ (coefficients in backward order).

Example:

k = 3
f_diff_weights(k)
4-element Vector{Int64}:
  1
 -3
  3
 -1

CamiXon.f_diff_weights_array — Method

f_diff_weights_array(kmax)

Collection of weight vectors $[c_k^k,\ \ldots,\ c_0^k]$ defining the finite difference operators $\nabla^0,\ \ldots,\ \nabla^k$.

Example:

kmax = 3
∇ = f_diff_weights_array(kmax)
4-element Vector{Vector{Int64}}:
 [1]
 [-1, 1]
 [1, -2, 1]
 [-1, 3, -3, 1]

Finite difference expansion

Finite-difference calculus builds on the finite-difference expansion, in practice truncated at order $k$,

\[\sum_{p=0}^{\infty}a_{p}\nabla^{p}f[n]=\sum_{p=0}^{k}a_{p}\nabla^{p}f[n]+\cdots.\]

By choosing the order sufficiently high the expansion can be approximated to any desired level of accuracy. A finite-difference expansion of order $k$ is defined by $k+1$ finite-difference expansion coefficients, above denoted by $a_{0},\ \ldots,\ a_{k}$. Note that both the coefficients and the operators appear in forward order. It takes some bookkeeping to rewrite the expansion in the form of a weighted sum over the $k+1$ function values in backward order $f[n],\ldots,f[n-k]$. Substituting the backward-difference expression for $\nabla^k$, we find

\[\sum_{p=0}^{k}a_{p}\nabla^{p}f[n]=\sum_{p=0}^{k}a_{p}\sum_{j=0}^{p}c_{j}^{i}f[n-j]=\sum_{j=0}^{k}\sum_{p=j}^{k}a_{p}c_{j}^{p}f[n-j]=\sum_{j=0}^{k}b_{j}^{k}f[n-j],\]

where the weighted summation is defined by the weights $b_{j}^{k}=\sum_{p=j}^{k}a_{p}c_{j}^{p}$, with $j=0,\ \ldots,\ k$. After changing dummy index to obtain forward order (coefficients in backward order) by the weights are generated by the function f_diff_expansion_weights(coeffs,∇), where ∇ =f_diff_weights_array(kmax).

Functions:

coeffs = $[a_0,\ \ldots,\ a_k]$

f_diff_expansion_weights(coeffs,∇) $\rightarrow [b_k^k,\ ,\ldots,\ b_0^k]$

CamiXon.f_diff_expansion_weights — Method

f_diff_expansion_weights(a, ∇)

Weight vector $[b_k^k,\ ,\ldots,\ b_0^k]$ corresponding to the expansion coefficients $[a_0^k,\ ,\ldots,\ a_k^k]$ of the $k^{th}$-order finite-difference expansion,

\[\sum_{p=0}^{k}a_{p}\nabla^{p}f[n]=\sum_{j=0}^{k}b_{j}^{k}f[n-j]=\sum_{j=0}^{k}b_{k-j}^{k}f[n-k+j],\]

where $f[n-k], ...,f[n]$ are elements of the analytic function $f$ tabulated in forward order. Note the difference in ordering between the finite-difference expansion coefficients, $a_{0},\ \ldots,\ a_{k}$, and the finite-difference expansion weights, $b_k^{k},\ \ldots,\ b_0^{k}$. Furthermore, the coefficients, $a_j$, are independent of $k$, whereas the weights, $b_j^k$, are $k$-dependent.

Example:

k=5
∇ = f_diff_weights_array(k)
a = UnitRange(0,k)
b = f_diff_expansion_weights(a, ∇)
6-element Vector{Int64}:
  -5
  29
 -69
  85
 -55
  15

Lagrangian interpolation

The definition of the finite difference can be rewritten in the form

\[(1-\nabla)f[n+1] = f[n].\]

Inverting this expression we have

\[f[n+1] = (1 - \nabla)^{-1} f[n] = \sum_{p=0}^{\infty}∇^p f[n],\]

\[f[n+2] = (1 - \nabla)^{-2} f[n] = \sum_{p=0}^{\infty}p∇^p f[n],\]

\[\vdots\]

which can be generalized into the form

\[f[n+x] = (1 - \nabla)^{-x} f[n] = \sum_{p=0}^{\infty}l_p(x)∇^p f[n],\]

where $l_0\equiv 1$ and $l_p(x) = x(x+1)(x+2)\cdots(x+p-1)/p!$ for $p>0$ are the finite-difference expansion coefficients for lagrangian interpolation. These are generated by the function f_diff_expansion_coeffs_interpolation(k,x). Evaluating the finite-difference expansion up to order $k$ we obtain (after changing dummy index to reverse the summation from backward to forward order)

\[f[n+x] =\sum_{p=0}^{k}l_p(x)\nabla^pf[n]= \sum_{j=0}^{k}r_j^k(x)f[n-j]= \sum_{j=0}^{k}r_{k-j}^k(x)f[n-k+j],\]

where the $k+1$ weights $r_j^k(x)= \sum_{p=j}^{k} l_p(x) c_j^p$ are the lagrangian interpolation weights. These are generated for use in backward order by the function f_diff_expansion_weights(coeffs,∇), with ∇ = f_diff_weights_array(k).

Functions:

coeffs = f_diff_expansion_coeffs_interpolation(k,x) $\rightarrow [l_0(x),\ \ldots,\ l_k(x)]$

lagrangian interpolation weights = f_diff_expansion_weights(coeffs,∇) $\rightarrow [r_k^k(x),\ ,\ldots,\ r_0^k(x)]$

CamiXon.f_diff_expansion_coeffs_interpolation — Method

f_diff_expansion_coeffs_interpolation(k::Int, x::T) where T<:Real

Finite-difference expansion coefficient vector $[l_0(x),\ \ldots,\ l_p(x)]$ defining $k^{th}$-order lagrangian interpolation of the tabulated analytic function $f(n+x)$ at offset position $x$ with respect to the position $n$, with $-k\le x\le 0$,

\[f[n+x] =\sum_{p=0}^{k}l_p(x)\nabla^pf[n],\]

where $l_0\equiv 1$ and $l_p(x) = x(x+1)(x+2)\cdots(x+p-1)/p!$.

Examples:

k=3
∇ = f_diff_weights_array(k)
x=-1
l = f_diff_expansion_coeffs_interpolation(k,x)
r = f_diff_expansion_weights(l, ∇)
println(l,r)
 [1, -1, 0, 0][0, 0, 1, 0]

CamiXon.f_diff_expansion_coeffs_extrapolation — Method

f_diff_expansion_coeffs_extrapolation(k::Int, x::T) where T<:Real

Finite-difference expansion coefficient vector $[l_0(x),\ \ldots,\ l_p(x)]$ defining $k^{th}$-order lagrangian extrapolation of the tabulated analytic function $f(n+x)$ at offset position $x$ with respect to the position $n$, with $x\le 0$,

\[f[n+x] =\sum_{p=0}^{k}l_p(x)\nabla^pf[n],\]

where $l_0\equiv 1$ and $l_p(x) = x(x+1)(x+2)\cdots(x+p-1)/p!$.

Examples:

k=5
∇ = f_diff_weights_array(k)
x=1
l = f_diff_expansion_coeffs_extrapolation(k,x)
r = f_diff_expansion_weights1(l, ∇)
println(l,r)
 [1, 1, 1, 1, 1, 1][-1, 6, -15, 20, -15, 6]

CamiXon.f_diff_expansion_coeffs_array_interpolation — Method

f_diff_expansion_coeffs_array_interpolation(k::Int, m::Int)

Finite-difference expansion coefficient vector $[l_0(x),\ \ldots,\ l_p(x)]$ defining $k^{th}$-order lagrangian interpolation of the tabulated analytic function $f(n+x)$ for $k*m+1$ values of $x$.

k = 2; m = 2
o = f_diff_expansion_coeffs_array_interpolation(k,m); println(o)
 [[1.0, 0.0, 0.0], [1.0, -0.5, -0.125], [1.0, -1.0, -0.0], [1.0, -1.5, 0.375], [1.0, -2.0, 1.0]]

m = 1
o = f_diff_expansion_coeffs_array_interpolation(k,m); println(o)
 [[1, 0, 0], [1, -1, 0], [1, -2, 1]]

CamiXon.f_diff_expansion_weights_extrapolation — Method

f_diff_expansion_weights_extrapolation(k::Int, x::T) where T<:Real

Finite-difference expansion weight vector for $k^{th}$-order lagrangian extrapolation of the tabulated analytic function $f(n+x)$ at offset position $x$ with respect to the position $n$, with $x\le 0$,

\[f[n+x] =\sum_{p=0}^{k}l_p(x)\nabla^pf[n],\]

where $l_0\equiv 1$ and $l_p(x) = x(x+1)(x+2)\cdots(x+p-1)/p!$.

Example:

k=5
∇ = f_diff_weights_array(k)
x=1
o = f_diff_expansion_weights_extrapolation(k,x); println(o)
 [-1, 6, -15, 20, -15, 6]

CamiXon.f_diff_expansion_weights_array — Method

f_diff_expansion_weights_array(n::Int, k::Int, m::Int, l::Vector{Vector{T}}) where T<:Real

Weight factors for $k^{th}$-order finite-difference expansion with finite-difference expansion coefficients $l$ on uniform grid of $(n-1)*m+1$ points.

Example:

n = 7; k = 3; m = 1
l = f_diff_expansion_coeffs_array_interpolation(k, m)
o = f_diff_expansion_weights_array(n, k, m, l); println(o)
 [[1, 0, 0, 0], [1, 0, 0, 0], [1, 0, 0, 0], [1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 1, 0], [0, 0, 0, 1]]

CamiXon.summation_ranges — Method

summation_ranges(n, k, i, m)

Summation ranges for interpolation positions $i = 0,\ 1,\ \ldots,\ (n-1)*m$ as used in $k^{th}$-order lagrangian interpolation of the anaytic function $f$ tabulated in forward order on a uniform grid of $n$ points, f[1], ...,f[n].

Examples:

n = 7; k = 2; m = 1
o = [summation_ranges(n,i,k,m) for i=0:(n-1)*m]; println(o)
 UnitRange{Int64}[1:3, 2:4, 3:5, 4:6, 5:7, 5:7, 5:7]

CamiXon.f_diff_function_sequences — Method

f_diff_function_sequences(f, n::Int, k::Int; i=0)

Finite-difference interpolation sequences of $k⋅m+1$ function values, where $m=i+1$, given in forward order including with $i=m-1$ intermediate points for use in$k^{th}$-order lagrangian intepolation of the anaytic function $f$ given in forward order at n points, $f[1], ...,f[n]$.

Example:

f = [0,1,2,3,4,5,6]
k = 2
o = f_diff_function_sequences(f, k); println(o)
 [[0, 1, 2], [1, 2, 3], [2, 3, 4], [3, 4, 5], [4, 5, 6], [4, 5, 6], [4, 5, 6]]

CamiXon.lagrangian_interpolation — Method

lagrangian_interpolation(f::Vector{Float64}, domain::ClosedInterval{Float64}; k=1, i=0)

$k^{th}$-order lagrangian-interpolation with $i$ intermediate point of the analytic function $f$ tabulated in forward order at n points, $f[1],\ \ldots,\ f[n]$.

Example:

f = [0.0,1,2,3,4,5,6,7]
domain = 0.0..1.0
(X,Y) = lagrangian_interpolation(f, domain; k=2, i=1); println((X,Y))
 (0.0:0.07142857142857142:1.0, [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0, 6.5, 7.0])

Lagrangian differentiation

To derive the lagrangian differentiation formulas we differentiate

\[f[n+x] = (1 - \nabla)^{-x} f[n]\]

analytically with respect to $x$.

\[\frac{df}{dx}[n+x]=-ln(1-\nabla)\ (1-\nabla)^{-x}f[n]=\sum_{q=1}^{k}\tfrac{1}{q}∇^{q}\sum_{p=0}^{k}l_{p}(x)∇^{p}f[n].\]

Expanding the r.h.s. in powers of $\nabla$ for given values of $n$ and $x$ we obtain an expression of the form

\[\frac{df}{dx}[n+x]=\sum_{p=1}^{k}l_{p}^{\prime}(x)\nabla^{p}f[n],\]

where $l_{p}^{\prime}(x)$ represents the finite-difference expansion coefficients for lagrangian differentiation at position $n+x$. These coefficients are determined numerically by polynomial multiplication after adding the $p=0$ term $l_{0}^{\prime}(x)\equiv 0$. These are generated by the function f_diff_expansion_coeffs_differentiation(k,x). Evaluating the finite-difference expansion up to order $k$ we obtain (after changing dummy index to reverse the summation from backward to forward order)

\[f[n+x]=\sum_{p=0}^{k}l_{p}^{\prime}(x)\nabla^{p}f[n]=\sum_{j=0}^{k}s_{j}^{k}(x)f[n-j]= \sum_{j=0}^{k}s_{k-j}^k(x)f[n-k+j],\]

where the $s_{j}^{k}(x)=\sum_{p=j}^{k}l_{p}^{\prime}(x)c_{j}^{p}$ are the $k$-point lagrangian differentiation weights. These are generated for use in backward order by the function f_diff_expansion_weights(coeffs,∇), with ∇ = f_diff_weights_array(k).

Functions:

coeffs = f_diff_expansion_coeffs_differentiation(k,x) $\rightarrow [l_0(x),\ \ldots,\ l_k(x)]$

lagrangian differentiation weights = f_diff_expansion_weights(coeffs,∇) $\rightarrow [s_k^k(x),\ ,\ldots,\ s_0^k(x)]$

CamiXon.f_diff_expansion_coeffs_differentiation — Method

f_diff_expansion_coeffs_differentiation(k::Int, x::T) where T<:Real

Finite-difference expansion coefficient vector $[l_0(x),\ \ldots,\ l_p(x)]$ defining $k^{th}$-order lagrangian differentiation of the tabulated analytic function $f(n+x)$ at position $x$.

Example:

k = 2; x = 0
o = f_diff_expansion_coeffs_differentiation(k,x); println(o)
 [0.0, 1.0, -1.5]

CamiXon.f_diff_expansion_coeffs_array_differentiation — Method

f_diff_expansion_coeffs_array_differentiation(k::Int, m::Int)

Finite-difference expansion coefficient vector $[l_0(x),\ \ldots,\ l_p(x)]$ defining $k^{th}$-order lagrangian differentiation of the tabulated analytic function $f(n+x)$ for $k\!m+1$ values of $x$ in the interval $[-k,\ \cdots,\ 0]$ and with $m-1$ intermediate points.

Examples:

k = 2; m = 2
o = f_diff_expansion_coeffs_array_differentiation(k,m); println(o)
 [[0.0, 1.0, 0.5], [0.0, 1.0, 0.0], [0.0, 1.0, -0.5], [0.0, 1.0, -1.0], [0.0, 1.0, -1.5]]

m = 1
o = f_diff_expansion_coeffs_array_differentiation(k,m); println(o)
 [[0.0, 1.0, 0.5], [0.0, 1.0, -0.5], [0.0, 1.0, -1.5]]

CamiXon.lagrangian_differentiation — Method

lagrangian_differentiation(f::Vector{Float64}, domain::ClosedInterval{Float64}; k=1, i=0)

$k^{th}$-order lagrangian differentiation with $i$ intermediate points of the function $f$ tabulated in forward order at n points, $f[1],\ \ldots,\ f[n]$.

Example:

f = [0.0,1,2,3,4,5]
domain = 0.0..5.0
X,Y = lagrangian_differentiation(f, domain; k=2, i = 0); println(X,Y)
 (0.0:1.0:5.0, [1.0, 1.0, 1.0, 1.0, 1.0, 1.0])

Adams-Bashford expansion

Adams-Moulton expansion

The Adams-Moulton integration step is given by the expansion

\[y[n+1]-y[n] = -\frac{h \nabla}{ln(1-\nabla)}f[n+1] = h\ ( 1 - \frac{1}{2}\nabla - \frac{1}{12}\nabla^2 - \frac{1}{24}\nabla^3 +\cdots)f[n+1].\]

For the evaluation of the integration step we limit the summation to $k+1$ terms (order $k$),

\[y[n+1]-y[n]= h\ (\sum_{p=0}^{k}b_p\nabla^p)f[n+1]+\cdots.\]

where $b_0,\ldots,b_k$ are the Adams-Moulton expansion coefficients, rational numbers generated numerically by the function f_diff_expansion_coeffs_adams_moulton(k). Extracting the greatest common denominator, $1/D$, the step becomes

\[y[n+1]-y[n]= \frac{h}{D}(\sum_{p=0}^{k}b_p^{\prime}\nabla^p)f[n+1]+\cdots,\]

where $b_0^{\prime},\ldots,b_k^{\prime}$ are integers and $b_p=b_p^{\prime}/D$. In practice the expansion is restricted to $k<18$ (as limited by integer overflow). Note that this limit is much higher than values used in calculations (typically up to $k = 10$).

Function:

coeffs = f_diff_expansion_coeffs_adams_moulton(k) $\rightarrow [b_k^k(x),\ ,\ldots,\ b_0^k(x)]$

CamiXon.f_diff_expansion_coeffs_adams_moulton — Method

f_diff_expansion_coeffs_adams_moulton(k::Int)

Examples:

k = 5
b = f_diff_expansion_coeffs_adams_moulton(k::Int); println(b)
 Rational[1//1, -1//2, -1//12, -1//24, -19//720, -3//160]

D = denominator(gcd(b)); println(D)
 1440

o = convert(Vector{Int},(b .* D)); println(o)
 [1440, -720, -120, -60, -38, -27]

FITS

FITS stands for 'Flexible Image Transport System'. This is an open standard origionally developed for the astronomy community to store telescope images together with tables of spectral information. Over the years it has developed into a scientific standard - http://fits.gsfc.nasa.gov/iaufwg.

Within CamiXion only the basic FITS functionality is implemented for users not requiring celestal coordinates. The user can create, read and extend .fits files as well as create, edit and delete user-defined metainformation.

A FITS file consists of a sequence of one or more header-data-units (HDUs), each containing a data block preceeded by header records of metainformation.

By the command f = fits_read(filnam) we asign a collection of FITS_HDU objects from the file filnam to the variable f.

FITS - Types

CamiXon.FITS_HDU — Type

FITS_HDU

Object to hold a single "Header-Data Unit" (HDU).

The fields are

.filename::String: name of the corresponding FITS file

.hduindex::Int: identifier (a file may contain more than one HDU)

.header::T, where T=FITS_header: the header object

.dataobject::V, where V=FITS_data: the data object

CamiXon.FITS_header — Type

FITS_header

Object to hold the header information of a FITS_HDU.

The fields are:

.hduindex::Int: identifier (a file may contain more than one HDU)

.records::Array{String,1}: the header formated as an array of strings of 80 ASCII characters

.keys::Array{String,1}: keys[i] - key corresponding to records[i] (record of index i)

.values::Array{Any,1}: value[i] - corresponding to records[i]

.comments: comments[i] - comment corresponding to records[i]

.dict::Dict{String,Any}: Dictionary key[i] => value[i]

.maps::Dict{String,Int}: Dictionary key[i] => i

CamiXon.FITS_data — Type

FITS_data

Object to hold the data of the FITS_HDU of given hduindex and hdutype.

The fields are:

.hduindex::Int: identifier (a file may contain more than one HDU)

.hdutype::String: accepted types are 'PRIMARY', 'IMAGE' and 'TABLE'

.data::Any: in the from appropriate for the hdutype

CamiXon.FITS_table — Type

FITS_table

Object to hold the data of a TABLE HDU (a FITS_HDU for ASCII tables). It contains the data in the form of records (rows) of ASCII strings.

The fields are:

.hduindex::Int: identifier (a file may contain more than one HDU)

.rows::Array{String,1}: the table formated as an array of rows of ASCII strings

CamiXon.FITS_name — Type

FITS_name

FITS object to decompose the names of .fits files.

The fields are:

.name::String: for filename 'p#.fits' this is 'p#.fits'

.prefix::String: for filename 'p#.fits' this is 'p'

.numerator::String: for filename 'p#.fits' this is '#', a serial number (e.g., '3') or a range (e.g., '3-7')

.extension::String: for filename 'p#.fits' this is '.fits'.

FITS - HDU Methods

CamiXon.fits_info — Method

fits_info(hdu)

Print metafinformation and data of given FITS_HDU

Example:

strExample = "remove.fits"
data = [11,21,31,12,22,23,13,23,33]
data = reshape(data,(3,3,1))
fits_create(strExample, data; protect=false)

f = fits_read(strExample)
fits_info(f[1])

  File: remove.fits
  hdu: 1
  hdutype: PRIMARY
  DataType: Int64
  Datasize: (3, 3, 1)

  Metainformation:
  SIMPLE  =                    T / file does conform to FITS standard
  BITPIX  =                   64 / number of bits per data pixel
  NAXIS   =                    3 / number of data axes
  NAXIS1  =                    3 / length of data axis 1
  NAXIS2  =                    3 / length of data axis 2
  NAXIS3  =                    1 / length of data axis 3
  BZERO   =                  0.0 / offset data range to that of unsigned integer
  BSCALE  =                  1.0 / default scaling factor
  EXTEND  =                    T / FITS dataset may contain extensions
  COMMENT    Primary FITS HDU    / http://fits.gsfc.nasa.gov/iaufwg
  END

  3×3×1 Array{Int64, 3}:
  [:, :, 1] =
   11  12  13
   21  22  23
   31  23  33

CamiXon.parse_FITS_TABLE — Method

parse_FITS_TABLE(hdu)

Parse FITS_TABLE (ASCII table) into a Vector of its columns for further processing by the user. Default formatting in ISO 2004 FORTRAN data format specified by keys "TFORMS1" - "TFORMSn") Display formatting in ISO 2004 FORTRAN data format ("TDISP1" - "TDISPn") prepared for user editing.

Example:

strExample = "example.fits"
data = [10, 20, 30]
fits_create(strExample, data; protect=false)

t1 = Float16[1.01E-6,2.0E-6,3.0E-6,4.0E-6,5.0E-6]
t2 = [0x0000043e, 0x0000040c, 0x0000041f, 0x0000042e, 0x0000042f]
t3 = [1.23,2.12,3.,4.,5.]
t4 = ['a','b','c','d','e']
t5 = ["a","bb","ccc","dddd","ABCeeaeeEEEEEEEEEEEE"]
data = [t1,t2,t3,t4,t5]
fits_extend(strExample, data, "TABLE")

f = fits_read(strExample)
d = f[2].header.dict
d = [get(d,"TFORM$i",0) for i=1:5]; println(strip.(d))
  SubString{String}["'E6.1    '", "'I4      '", "'F4.2    '", "'A1      '", "'A20     '"]

f[2].dataobject.data                            # this is the table hdu
  5-element Vector{String}:
   "1.0e-6 1086 1.23 a a                    "
   "2.0e-6 1036 2.12 b bb                   "
   "3.0e-6 1055 3.0  c ccc                  "
   "4.0e-6 1070 4.0  d dddd                 "
   "5.0e-6 1071 5.0  e ABCeeaeeEEEEEEEEEEEE "

parse_FITS_TABLE(f[2])
  5-element Vector{Vector{T} where T}:
   [1.0e-6, 2.0e-6, 3.0e-6, 4.0e-6, 5.0e-6]
   [1086, 1036, 1055, 1070, 1071]
   [1.23, 2.12, 3.0, 4.0, 5.0]
   ["a", "b", "c", "d", "e"]
   ["a                   ", "bb                  ", "ccc                 ", "dddd                ", "ABCeeaeeEEEEEEEEEEEE"]

FITS - File Methods

CamiXon.cast_FITS_name — Method

cast_FITS_name(filename::String)

Decompose the FITS filename 'filnam.fits' into its name, prefix, numerator and extension.

Examples:

strExample = "T23.01.fits"
f = cast_FITS_name(strExample)
FITS_name("T23.01", "T23.", "01", ".fits")

f.name, f.prefix, f.numerator, f.extension
("T23.01", "T23.", "01", ".fits")

CamiXon.fits_combine — Method

fits_combine(strFirst, strLast [; protect=true])

Copy "filenameFirst" to "filenameLast" (with mandatory ".fits" extension)

Key:

protect::Bool: overwrite protection

Example:

fits_combine("T01.fits", "T22.fits")
  'T01-T22.fits': file created

CamiXon.fits_copy — Function

fits_copy(filenameA [, filenameB="" [; protect=true]])

Copy "filenameA" to "filenameB" (with mandatory ".fits" extension) Key:

protect::Bool: overwrite protection

Examples:

fits_copy("T01.fits")
  'T01.fits' was saved as 'T01 - Copy.fits'

fits_copy("T01.fits", "T01a.fits")
  FitsError: 'T01a.fits' in use (set ';protect=false' to lift overwrite protection)

fits_copy("T01.fits", "T01a.fits"; protect=false)
  'T01.fits' was saved as 'T01a.fits'

CamiXon.fits_create — Function

fits_create(filename [, data [; protect=true]])

Create FITS file of given filename [, optional data block [, default overwrite protection]] and return Array of HDUs. Key:

protect::Bool: overwrite protection

Examples:

strExample = "minimal.fits"
fits_create(strExample;protect=false)

f = fits_read(strExample)
a = f[1].dataobject.data
b = f[1].header.keys
println(a);println(b)
  Any[]
  ["SIMPLE", "NAXIS", "EXTEND", "COMMENT", "END"]

strExample = "remove.fits"
data = [11,21,31,12,22,23,13,23,33]
data = reshape(data,(3,3,1))
fits_create(strExample, data; protect=false)

f = fits_read(strExample)
fits_info(f[1])

  File: remove.fits
  hdu: 1
  hdutype: PRIMARY
  DataType: Int64
  Datasize: (3, 3, 1)

  Metainformation:
  SIMPLE  =                    T / file does conform to FITS standard
  BITPIX  =                   64 / number of bits per data pixel
  NAXIS   =                    3 / number of data axes
  NAXIS1  =                    3 / length of data axis 1
  NAXIS2  =                    3 / length of data axis 2
  NAXIS3  =                    1 / length of data axis 3
  BZERO   =                  0.0 / offset data range to that of unsigned integer
  BSCALE  =                  1.0 / default scaling factor
  EXTEND  =                    T / FITS dataset may contain extensions
  COMMENT    Primary FITS HDU    / http://fits.gsfc.nasa.gov/iaufwg
  END

  3×3×1 Array{Int64, 3}:
  [:, :, 1] =
   11  12  13
   21  22  23
   31  23  33

CamiXon.fits_extend — Function

fits_extend(filename, data_extend, hdutype="IMAGE")

Extend the FITS file of given filename with the data of hdutype from data_extend and return Array of HDUs.

Examples:

strExample = "test_example.fits"
data = [0x0000043e, 0x0000040c, 0x0000041f]
fits_create(strExample, data, protect=false)

f = fits_read(strExample)
a = Float16[1.01E-6,2.0E-6,3.0E-6,4.0E-6,5.0E-6]
b = [0x0000043e, 0x0000040c, 0x0000041f, 0x0000042e, 0x0000042f]
c = [1.23,2.12,3.,4.,5.]
d = ['a','b','c','d','e']
e = ["a","bb","ccc","dddd","ABCeeaeeEEEEEEEEEEEE"]
data = [a,b,c,d,e]
fits_extend(strExample, data, "TABLE")

f = fits_read(strExample)
f[2].dataobject.data
  5-element Vector{String}:
   "1.0e-6 1086 1.23 a a                    "
   "2.0e-6 1036 2.12 b bb                   "
   "3.0e-6 1055 3.0  c ccc                  "
   "4.0e-6 1070 4.0  d dddd                 "
   "5.0e-6 1071 5.0  e ABCeeaeeEEEEEEEEEEEE "

rm(strExample); f = data = a = b = c = d = e = nothing

CamiXon.fits_read — Method

fits_read(filename)

Read FITS file and return Array of FITS_HDUs

Example:

strExample = "minimal.fits"
fits_create(strExample;protect=false)

f = fits_read(strExample)
f[1].dataobject.data
  Any[]

rm(strExample); f = nothing

FITS - Key Methods

CamiXon.fits_add_key — Method

fits_add_key(filename, hduindex, key, value, comment)

Add a header record of given 'key, value and comment' to 'HDU[hduindex]' of file with name 'filename'

Example:

strExample="minimal.fits"
fits_create(strExample;protect=false)
fits_add_key(strExample, 1, "KEYNEW1", true, "FITS dataset may contain extension")

f = fits_read(strExample)
fits_info(f[1])

  File: minimal.fits
  hdu: 1
  hdutype: PRIMARY
  DataType: Any
  Datasize: (0,)

  Metainformation:
  SIMPLE  =                    T / file does conform to FITS standard
  NAXIS   =                    0 / number of data axes
  EXTEND  =                    T / FITS dataset may contain extensions
  COMMENT    Primary FITS HDU    / http://fits.gsfc.nasa.gov/iaufwg
  KEYNEW1 =                    T / FITS dataset may contain extension
  END

  Any[]

CamiXon.fits_delete_key — Method

fits_delete_key(filename, hduindex, key)

Delete a header record of given key, value and comment to FITS_HDU[hduindex] of file with name 'filename'

Examples:

strExample="minimal.fits"
fits_create(strExample;protect=false)
fits_add_key(strExample, 1, "KEYNEW1", true, "this is record 5")

f = fits_read(strExample)
get(f[1].header.maps,"KEYNEW1",0)
  5

fits_delete_key(strExample, 1, "KEYNEW1")

f = fits_read(strExample)
get(f[1].header.maps,"KEYNEW1",0)
  0

fits_delete_key(filnam, 1, "NAXIS")
 'NAXIS': cannot be deleted (key protected under FITS standard)

CamiXon.fits_edit_key — Method

fits_edit_key(filename, hduindex, key, value, comment)

Edit a header record of given 'key, value and comment' to 'HDU[hduindex]' of file with name 'filename'

Example:

data = DateTime("2020-01-01T00:00:00.000")
strExample="minimal.fits"
fits_create(strExample;protect=false)
fits_add_key(strExample, 1, "KEYNEW1", true, "this is record 5")
fits_edit_key(strExample, 1, "KEYNEW1", data, "record 5 changed to a DateTime type")

f = fits_read(strExample)
fits_info(f[1])

  File: minimal.fits
  hdu: 1
  hdutype: PRIMARY
  DataType: Any
  Datasize: (0,)

  Metainformation:
  SIMPLE  =                    T / file does conform to FITS standard
  NAXIS   =                    0 / number of data axes
  EXTEND  =                    T / FITS dataset may contain extensions
  COMMENT    Primary FITS HDU    / http://fits.gsfc.nasa.gov/iaufwg
  KEYNEW1 = '2020-01-01T00:00:00' / record 5 changed to a DateTime type
  END

  Any[]

CamiXon.fits_rename_key — Method

fits_rename_key(filename, hduindex, keyold, kewnew)

Rename the key of a header record of file with name 'filename'

Example:

strExample="minimal.fits"
fits_create(strExample;protect=false)
fits_add_key(strExample, 1, "KEYNEW1", true, "this is record 5")
fits_rename_key(strExample, 1, "KEYNEW1",  "KEYNEW2")

f = fits_read(strExample)
fits_info(f[1])

  File: minimal.fits
  hdu: 1
  hdutype: PRIMARY
  DataType: Any
  Datasize: (0,)

  Metainformation:
  SIMPLE  =                    T / file does conform to FITS standard
  NAXIS   =                    0 / number of data axes
  EXTEND  =                    T / FITS dataset may contain extensions
  COMMENT    Primary FITS HDU    / http://fits.gsfc.nasa.gov/iaufwg
  KEYNEW2 =                    T / this is record 5
  END

  Any[]

FORTRAN

CamiXon.FORTRAN_format — Type

FORTRAN_format

Object to hold a FORTRAN format specifier decomposed into its fields. Accepted datatype specifiers are: Aw, Iw, Fw.d, Ew.d, Dw.d. Accepted output formating specifiers are: Aw, Iw.m, Bw.m, Ow.m, Zw.m, Fw.d, Ew.dEe, ENw.d, ESw.d, Gw.dEe, Dw.dEe. Notation: 'w' - width, 'm' (optional) - minimum number of digits, 'd' - number of digits to right of decimal, 'e' - number of digits in exponent; 'N'/'S' (optional) - switch for engineering/scientific formating of the 'E' type.

The fields are:

Type::String: primary FORTRAN datatype

TypeChar::Char: primary FORTRAN datatype character

EngSci::Union{Char,Nothing}: secundary datatype character (N for engineering, S for scientific)

width::Int: width of numeric field

nmin::Int: minimum number of digits displayed

ndec::Int: number of digits to right of decimal

nexp::Int: number of digits in exponent

CamiXon.cast_FORTRAN_format — Method

cast_FORTRAN_format(format::String)

Decompose the format specifier 'format' into its fields and cast this into the FORTRAN_format object. Allowed format specifiers are of the types: Aw, Iw.m, Bw.m, Ow.m, Zw.m, Fw.d, Ew.dEe, ENw.d, ESw.d, Gw.dEe, Dw.dEe, with: 'w' - width, 'm' (optional) - minimum number of digits, 'd' - number of digits to right of decimal, 'e' - number of digits in exponent; 'N'/'S' (optional) - switch for engineering/scientific formating of the 'E' type.

Examples:

t = cast_FORTRAN_format("I10")
FORTRAN_format("Iw", 'I', nothing, 10, 0, 0, 0)

t = cast_FORTRAN_format("I10.12")
FORTRAN_format("Iw.m", 'I', nothing, 10, 12, 0, 0)

t = cast_FORTRAN_format("E10.5E3")
FORTRAN_format("Ew.dEe", 'E', nothing, 10, 0, 5, 3)

t.Type, t.TypeChar, t.EngSci, t.width, t.nmin, t.ndec, t.nexp
("Ew.dEe", 'E', nothing, 10, 0, 5, 3)

CamiXon.cast_FORTRAN_datatype — Method

cast_FORTRAN_datatype(format::String)

Decompose the format specifier 'format' into its fields and cast this into the FORTRAN_format object. Allowed format specifiers are of the types: Aw, Iw, Fw.d, Ew.d, Dw.d, where: 'w' - width, 'd' - number of digits to right of decimal point.

Examples:

f = cast_FORTRAN_datatype("I10")
FORTRAN_format("Iw", 'I', nothing, 10, 0, 0, 0)

f = cast_FORTRAN_datatypet("F10.4")
FORTRAN_format("Fw.d", 'F', nothing, 10, 0, 4, 0)

f = cast_FORTRAN_datatype("E10.5")
FORTRAN_format("Ew.d", 'E', nothing, 10, 0, 5, 0)

t.Type, t.TypeChar, t.EngSci, t.width, t.nmin, t.ndec, t.nexp
("Ew.d", 'E', nothing, 10, 0, 5, 0)

Plotting

CamiXon.step125 — Method

step125(x)

Step used for deviding the number x in steps according to 1-2-5 scheme

Examples:

step125.([5,10,21.3,50,100.1])
5-element Vector{Int64}:
  1
  2
  5
 10
 20

CamiXon.select125 — Method

select125(x)

Select elements of the collection x by index according to 1-2-5 scheme

Examples:

x = [1,2,4,6,8,10,13,16,18,20,40,60,80,100]
select125(x)
 [2, 6, 10, 16, 20, 60, 100]

x = string.(x)
select125(x)
 ["2", "6", "10", "16", "20", "60", "100"]

x = 1:100
select125(x)
 [20, 40, 60, 80, 100]

CamiXon.steps — Method

steps(x)

Heatmap range transformation for steplength specification vector x

Examples:

x = [4,2,6]
steps(x)
 [0, 4, 6, 12]

CamiXon.stepcenters — Method

stepcenters(x)

Stepcenter positions for steplength specification vector x

Examples:

x = [4,2,6]
stepcenters(x)
 [2.0, 5.0, 9.0]

CamiXon.stepedges — Method

stepedges(x)

Stepedges for steplength specification vector x

Examples:

x = [4,2,6]
stepedges(x)
 [0, 4, 6, 12]

CamiXon.edges — Function

edges(px [, Δx[, x0]])

Heatmap range transformation from pixel coordinates to physical coordinates, with pixelsize Δx and offset x0, both in physical units.

Examples:

px = 1:5
Δx = 2.5
x0 = 2.5
edges(px)
 [0.5, 1.5, 2.5, 3.5, 4.5]

edges(px, Δx)
 [1.25, 3.75, 6.25, 8.75, 11.25]

edges(px, Δx, x0)
 [-1.25, 1.25, 3.75, 6.25, 8.75]

Search algorithms

CamiXon.find_all — Method

find_all(A [,a...]; count=false)

A: string/array of elements of the same type

default : Array containing the index (indices) of selected elements of A (default: all elements)

count=true: The number of indices found for selected elements of A (default: all elements)

Examples:

A = [:📑,:📌,:📢,:📌,:📞]
B = [1,2,3,2,5]
str = "aβcβd";
find_all(A) == find_all(B) == find_all(str)
true

find_all(A,:📌)
1-element Array{Array{Int64,1},1}:
 [2, 4]

find_all(str)
4-element Array{Array{Int64,1},1}:
 [1]
 [2, 4]
 [3]
 [5]

find_all(A; count=true)
4-element Array{Int64,1}:
 1
 2
 1
 1

str = "📑📌📢📌📞"
find_all(str,'📌')
1-element Array{Array{Int64,1},1}:
 [2, 4]

CamiXon.find_first — Method

find_first(A [,a...]; dict=false)

The first index of selected Array element

A: string/array of elements of the same type

default : Array containing the first index (indices) of selected elements of A (default: all elements)

dict=true: Dict for the first index (indices) of selected elements of A (default: all elements)

Examples:

A = [:📑,:📌,:📢,:📌,:📞]
B = [1,2,3,2,5]
str = "aβcβd";

find_first(A) == find_first(B) == find_first(str)
true

find_first(A,:📌)
1-element Array{Array{Int64,1},1}:
 2

find_last(A,:📌; dict=true)
1-element Array{Pair{Symbol,Int64},1}:
 :📌 => 2

find_last(A; dict=true)
4-element Array{Pair{Symbol,Int64},1}:
 :📑 => 1
 :📌 => 2
 :📢 => 3
 :📞 => 5

find_first(str)
4-element Array{Int64,1}:
 1
 2
 3
 5

CamiXon.find_last — Method

find_last(A [,a...]; dict=false)

The last index of selected Array element

A: string/array of elements of the same type

default : Array containing the lasst index (indices) of selected elements of A (default: all elements)

dict=true: Dict for the lasst index (indices) of selected elements of A (default: all elements)

Examples:

A = [:📑,:📌,:📢,:📌,:📞]
B = [1,2,3,2,5]
str = "aβcβd";
find_last(A) == find_first(B) == find_first(str)
true

find_last(A,:📌)
1-element Array{Array{Int64,1},1}:
 4

find_last(A,:📌; dict=true)
1-element Array{Pair{Symbol,Int64},1}:
 :📌 => 4

find_last(A; dict=true)
4-element Array{Pair{Symbol,Int64},1}:
 :📑 => 1
 :📌 => 4
 :📢 => 3
 :📞 => 5

find_last(str)
4-element Array{Int64,1}:
 1
 4
 3
 5

Math

CamiXon.canonical_partitions — Function

canonical_partitions(n; header=false, reverse=true)

The canonical partition in integers of the integer n

header=true : unit patition included in output

Examples:

canonical_partitions(6; header=true, reverse=false)
6-element Array{Array{Int64,1},1}:
 [6]
 [5, 1]
 [4, 2]
 [3, 3]
 [2, 2, 2]
 [1, 1, 1, 1, 1, 1]

canonical_partitions(6; header=true)
6-element Array{Array{Int64,1},1}:
 [1, 1, 1, 1, 1, 1]
 [2, 2, 2]
 [3, 3]
 [4, 2]
 [5, 1]
 [6]

canonical_partitions(6)
5-element Array{Array{Int64,1},1}:
 [1, 1, 1, 1, 1, 1]
 [2, 2, 2]
 [3, 3]
 [4, 2]
 [5, 1]

CamiXon.integer_partitions — Function

integer_partitions(n [,m]; transpose=false, count=false)

default : The integer partitions of n

count=true : The number of integer partitions of n

transpose=false/true : for m>0 restricted to partitions with maximum part/length m

definitions:

The integer partition of the positive integer n is a nonincreasing sequence of positive integers p1, p2,... pk whose sum is n.

The elements of the sequence are called the parts of the partition.

Examples:

integer_partitions(7)
15-element Array{Array{Int64,1},1}:
 [1, 1, 1, 1, 1, 1, 1]
 [2, 2, 2, 1]
 [3, 3, 1]
 [4, 3]
 [5, 2]
 [6, 1]
 [7]
 [2, 2, 1, 1, 1]
 [3, 2, 2]
 [4, 2, 1]
 [5, 1, 1]
 [2, 1, 1, 1, 1, 1]
 [3, 2, 1, 1]
 [4, 1, 1, 1]
 [3, 1, 1, 1, 1]

integer_partitions(7; count=true)
15

integer_partitions(7,4; count=true)
3

integer_partitions(7,4)
3-element Array{Array{Int64,1},1}:
 [4, 3]
 [4, 2, 1]
 [4, 1, 1, 1]

integer_partitions(7,4; transpose=true)
3-element Array{Array{Int64,1},1}:
 [2, 2, 2, 1]
 [3, 2, 1, 1]
 [4, 1, 1, 1]

CamiXon.log10_characteristic_power — Method

log10_characteristic_power(x)

characteristic power-of-10 of the number x

Examples:

log10_characteristic_power.([3,30,300])
3-element Vector{Int64}:
 0
 1
 2

CamiXon.log10_mantissa — Method

log10_mantissa(x)

log10 mantissa of the number x

Examples:

log10_mantissa.([3,30,300])
3-element Vector{Float64}:
 0.47712125471966244
 0.4771212547196624
 0.4771212547196626

CamiXon.polynom_deriv_coeffs — Function

polynom_deriv_coeffs(c[,deriv=0])

Coefficient vector defining the derivative deriv of the polynomial $p(c,d)$, of degree $d$, defined by the coefficient vector c$=[c_0 +\ \ldots,\ c_d]$.

Examples:

d = 5
c = [1.0 for i=1:d+1]
polynom_deriv_coeffs(c)                # default is direct copy of `c`
 [1.0, 1.0, 1.0, 1.0, 1.0, 1.0]
polynom_deriv_coeffs(c,1)              # coefficients of 1st derivative of `polynom(c,x)`
 [0.0, 1.0, 2.0, 3.0, 4.0, 5.0]
polynom_deriv_coeffs(c,2)              # coefficients of 2nd derivative of `polynom(c,x)`
 [-0.0, 0.0, 2.0, 6.0, 12.0, 20.0]
polynom_deriv_coeffs(c,5)              # coefficients of 5th derivative of `polynom(c,x)`
 [0.0, -0.0, 0.0, -0.0, 0.0, 120.0]
polynom_deriv_coeffs(c,6)              # coefficients of 6th derivative of `polynom(c,x)`
 [0.0, 0.0, 0.0, 0.0, 0.0, 0.0]

CamiXon.polynom — Method

polynom(c,x)

Coefficient vector $c=[c_0 +\ \ldots,\ c_d]$ defining the polynomial of degree $d$,

\[p(c,x)=c_0 + c_1*x +\ \cdots,\c_n*x^d],\]

where $d = length(c)-1$.

Examples:

d = 5
c = [1.0 for i=1:d+1]
c = polynom_deriv_coeffs(c)          # default is simple
println(c)
 [1.0, 1.0, 1.0, 1.0, 1.0, 1.0]
f(x) = polynom(c,x)
println([f(1.0),f(2.0)])             # values of polynomial for x = 1.0 and x = 2.0
 [6.0, 63.0]

CamiXon.polynom_multiplication_coeffs — Method

polynom_multiplication_coeffs(a::Vector{<:Number}, b::Vector{<:Number})

Coefficient vector $c=[c_0,\ \ldots,\ c_{n+m}]$ of the polynomial of degree $m+n$

\[ p(c,x)=a_0b_0 + (a_0b_1 + b_0a_1)x +\ \cdots\ a_n b_m x^{n+m},\]

given by the product of two polynomials, of degree $n$ and $m$, defined by the coefficient vectors $a=[a_0,\ \ldots,\ a_n]$ and $b=[b_0,\ \ldots,\ b_m]$.

a = [1,1]
b = [1,- 1]
o = polynomial_multiplication_coeffs(a, b); println(o)
 [1, 0, -1]

CamiXon.permutations_unique_count — Method

permutations_unique_count(p::Array{Array{Int64,1},1}, i::Int)

Number of unique permutations of the subarray $p[i]$.

Example:

p = [[1,2,3],[2,3,1,4,3]]
permutations_unique_count(p,2)
 60

Index

CamiXon.FITS_HDU

CamiXon.FITS_data

CamiXon.FITS_header

CamiXon.FITS_name

CamiXon.FITS_table

CamiXon.FORTRAN_format

CamiXon.canonical_partitions

CamiXon.cast_FITS_name

CamiXon.cast_FORTRAN_datatype

CamiXon.cast_FORTRAN_format

CamiXon.edges

CamiXon.f_diff_expansion_coeffs_adams_moulton

CamiXon.f_diff_expansion_coeffs_array_differentiation

CamiXon.f_diff_expansion_coeffs_array_interpolation

CamiXon.f_diff_expansion_coeffs_differentiation

CamiXon.f_diff_expansion_coeffs_extrapolation

CamiXon.f_diff_expansion_coeffs_interpolation

CamiXon.f_diff_expansion_weights

CamiXon.f_diff_expansion_weights_array

CamiXon.f_diff_expansion_weights_extrapolation

CamiXon.f_diff_function_sequences

CamiXon.f_diff_weight

CamiXon.f_diff_weights

CamiXon.f_diff_weights_array

CamiXon.find_all

CamiXon.find_first

CamiXon.find_last

CamiXon.fits_add_key

CamiXon.fits_combine

CamiXon.fits_copy

CamiXon.fits_create

CamiXon.fits_delete_key

CamiXon.fits_edit_key

CamiXon.fits_extend

CamiXon.fits_info

CamiXon.fits_read

CamiXon.fits_rename_key

CamiXon.integer_partitions

CamiXon.lagrangian_differentiation

CamiXon.lagrangian_interpolation

CamiXon.log10_characteristic_power

CamiXon.log10_mantissa

CamiXon.parse_FITS_TABLE

CamiXon.permutations_unique_count

CamiXon.polynom

CamiXon.polynom_deriv_coeffs

CamiXon.polynom_multiplication_coeffs

CamiXon.select125

CamiXon.step125

CamiXon.stepcenters

CamiXon.stepedges

CamiXon.steps

CamiXon.summation_ranges