Orbital properties

Orbital

CamiXon.Orbit — Type

Orbit(name, n, n′, ℓ, mℓ)

Type for specification of atomic orbitals with fields:

.name: name

.n: principal quantum number

.n′: radial quantum number (number of nodes in radial wavefunction)

.ℓ: orbital angular momentum valence electron

.mℓ: orbital angular momentum projection valence electron

The type Orbit is best created with the function castOrbit.

CamiXon.castOrbit — Method

castOrbit(;n=1, ℓ=0, mℓ=0, msg=true)

Create Orbit with fields:

.name: name

.n: principal quantum number

.n′: radial quantum number (number of nodes in radial wavefunction)

.ℓ: orbital angular momentum valence electron

.mℓ: orbital angular momentum projection valence electron

Examples:

julia> castOrbit(n=1, ℓ=0; msg=true)
Orbital: 1s
    principal quantum number: n = 1
    radial quantum number: n′ = 0 (number of nodes in radial wavefunction)
    orbital angular momentum of valence electron: ℓ = 0
    orbital angular momentum projection of valence electron: mℓ = 0
Orbit("1s", 1, 0, 0, 0)

Spinorbital

CamiXon.Spinorbit — Type

Spinorbit

Type for specification of atomic Spinorbitals with fields:

.name: spinorbital name (string)

.orbit: orbital object (Orbit)

.ms: spin magnetic quantum number (Rational{Int})

The type Spinorbit is best created with the function castSpinorbit.

CamiXon.castSpinorbit — Method

castSpinorbit(;n=1, ℓ=0, mℓ=0, up=true, msg=false)

Create Spinorbit with fields:

.name: spinorbital name (string)

.orbit: orbital object (Orbit)

.ms: spin magnetic quantum number (Rational{Int})

Example:

julia> castSpinorbit(n=1, ℓ=0, msg=true)
Spinorbital: 1s↑
    principal quantum number: n = 1
    radial quantum number: n′ = 0 (number of nodes in radial wavefunction)
    orbital angular momentum of valence electron: ℓ = 0
    orbital angular momentum projection of valence electron: mℓ = 0
    spin magnetic quantum number: ms = 1/2
Spinorbit("1s↑", Orbit("1s", 1, 0, 0, 0), 1//2)

Term

CamiXon.Term — Type

Term(name::String, n::Int, ℓ::Int, S::Real, L::Int, J::Real)

Type for specification of atomic fine-structure Terms with fields:

name: name

.n: principal quantum number

.n′: radial quantum number (number of nodes in wavefunction)

.ℓ: orbital angular momentum valence electron

.S: total electron spin in units of ħ

.L: total orbital angular momentum in units of ħ

.J: total electronic angular momentum in units of ħ

The type Term is best created with the function castTerm.

CamiXon.castTerm — Method

castTerm(n::Int; ℓ=0, S=1//2, L=0, J=1//2, msg=false)

Specify Term in the Term notatation with fields:

.n: principal quantum number

.n′: radial quantum number (number of nodes - autogenerated)

.ℓ: orbital angular momentum valence electron

.S: total electron spin

.L: total orbital angular momentum

.J: total electronic angular momentum

Examples:

julia> castTerm(1; ℓ=0, S=1//2, L=0, J=1//2, msg=true)
Term created: 1s ²S₁⸝₂; n = 1,  n′ = 0, ℓ = 0, S = 1//2, L = 0, J = 1//2
Term("1s ²S₁⸝₂", 1, 0, 0, 1//2, 0, 1//2)